Module openmls::ciphersuite

Ciphersuites for MLS

This file contains the API to interact with ciphersuites. See codec.rs and ciphersuites.rs for internals.

Modules

ciphersuites

Ciphersuite and algorithm definitions and conversions.

hash_ref

Hash References

signable

This module defines traits used for signing and verifying structs from the MLS protocol spec.

Structs

Ciphersuite

The Ciphersuite object encapsulates all the necessary crypto primitives for a given CiphersuiteName.