Module openmls::config

This config contains all structs, enums and functions to configure MLS.

Re-exports

pub use codec::*;

pub use errors::ConfigError;

Modules

codec

errors
Config errors

Structs

CONFIG

Config
OpenMLS Configuration

Constants
Constants that are used throughout the library.

PersistentConfig
The configuration we use for the library (Config) is not exactly the same as the one we persist.

Enums

ProtocolVersion
Protocol Version

Constants

DEFAULT_KEY_PACKAGE_LIFETIME
This value is used as the default lifetime of KeyPackages if no default lifetime is configured. The value is in seconds and amounts to 3 * 28 Days, i.e. about 3 months.

DEFAULT_KEY_PACKAGE_LIFETIME_MARGIN
This value is used as the default amount of time (in seconds) the lifetime of a KeyPackage is extended into the past to allow for skewed clocks. The value is in seconds and amounts to 1h.

SUPPORTED_CIPHERSUITE_NAMES
Supported ciphersuites TODO #13: This should come from the crypto provider