Struct openmls::credentials::CredentialBundle

pub struct CredentialBundle { /* private fields */ }

Credential Bundle.

This struct contains a Credential and the private key corresponding to the signature key it contains.

Implementations

impl CredentialBundle

pub fn new(
identity: Vec<u8>,
credential_type: CredentialType,
signature_scheme: SignatureScheme,
backend: &impl OpenMlsCryptoProvider
) -> Result<Self, CredentialError>

Creates and returns a new CredentialBundle of the given CredentialType for the given identity and SignatureScheme. The corresponding key material is freshly generated.

Returns an error if the given CredentialType is not supported.

pub fn from_parts(identity: Vec<u8>, keypair: SignatureKeypair) -> Self

Creates a new CredentialBundle from an identity and a SignatureKeypair. Note that only BasicCredential is currently supported.

pub fn credential(&self) -> &Credential

Returns a reference to the Credential.

pub fn into_parts(self) -> (Credential, SignaturePrivateKey)

Separates the bundle into the Credential and the SignaturePrivateKey.