openmls::credentials::test_utils

Function new_credential

pub fn new_credential(
    provider: &impl OpenMlsProvider,
    identity: &[u8],
    signature_scheme: SignatureScheme,
) -> (CredentialWithKey, SignatureKeyPair)

Available on (32-bit or 64-bit) and (crate features test-utils) only.

Convenience function that generates a new credential and a key pair for it (using the basic credential crate). The signature keys are stored in the key store.

Returns the Credential and the [SignatureKeyPair].