openmls::error
Struct LibraryError

pub struct LibraryError { /* private fields */ }

Generic error type that indicates unrecoverable errors in the library.

This error has 3 subtypes:

MissingBoundsCheck

This error is returned when the library tries to serialize data that is too big for the MLS structs. In particular, when element lists contain more elements than the theoretical maximum defined in the spec, the serialization will fail. This should not happen when all input values are checked.

CryptoError

This error is returned if the underlying crypto provider encountered an unexpected error. Possible reasons for this could be: the implementation of the crypto provider is not correct, the key material is not correct, the crypto provider does not support all functions required. Another reason could be that the OpenMLS library does not use the crypto provider API correctly.

Custom

This error is returned in situations where the implementation would otherwise use an unwrap(). If applications receive this error, it clearly indicates an implementation mistake in OpenMLS. The error includes a string that can give some more context about where the error originated and helps debugging.

In all cases, when a LibraryError is returned, applications should try to recover gracefully from it. It is recommended to log the error for potential debugging.