Enum openmls::extensions::Extension

pub enum Extension {
    ApplicationId(ApplicationIdExtension),
    RatchetTree(RatchetTreeExtension),
    RequiredCapabilities(RequiredCapabilitiesExtension),
    ExternalPub(ExternalPubExtension),
    ExternalSenders(ExternalSendersExtension),
    LastResort(LastResortExtension),
    Unknown(u16, UnknownExtension),
}

Extension

An extension is one of the Extension enum values. The enum provides a set of common functionality for all extensions.

See the individual extensions for more details on each extension.

// draft-ietf-mls-protocol-16
struct {
    ExtensionType extension_type;
    opaque extension_data<V>;
} Extension;

Variants

ApplicationId(ApplicationIdExtension)

An ApplicationIdExtension

RatchetTree(RatchetTreeExtension)

A RatchetTreeExtension

RequiredCapabilities(RequiredCapabilitiesExtension)

A RequiredCapabilitiesExtension

ExternalPub(ExternalPubExtension)

An ExternalPubExtension

ExternalSenders(ExternalSendersExtension)

An ExternalSendersExtension

LastResort(LastResortExtension)

A LastResortExtension

Unknown(u16, UnknownExtension)

A currently unknown extension.

Implementations

impl Extension

pub fn as_application_id_extension( &self ) -> Result<&ApplicationIdExtension, ExtensionError>

Get a reference to this extension as ApplicationIdExtension. Returns an ExtensionError::InvalidExtensionType if called on an Extension that’s not an ApplicationIdExtension.

pub fn as_ratchet_tree_extension( &self ) -> Result<&RatchetTreeExtension, ExtensionError>

Get a reference to this extension as RatchetTreeExtension. Returns an ExtensionError::InvalidExtensionType if called on an Extension that’s not a RatchetTreeExtension.

pub fn as_required_capabilities_extension( &self ) -> Result<&RequiredCapabilitiesExtension, ExtensionError>

Get a reference to this extension as RequiredCapabilitiesExtension. Returns an ExtensionError::InvalidExtensionType error if called on an Extension that’s not a RequiredCapabilitiesExtension.

pub fn as_external_pub_extension( &self ) -> Result<&ExternalPubExtension, ExtensionError>

Get a reference to this extension as ExternalPubExtension. Returns an ExtensionError::InvalidExtensionType error if called on an Extension that’s not a ExternalPubExtension.

pub fn as_external_senders_extension( &self ) -> Result<&ExternalSendersExtension, ExtensionError>

Get a reference to this extension as ExternalSendersExtension. Returns an ExtensionError::InvalidExtensionType error if called on an Extension that’s not a ExternalSendersExtension.

pub const fn extension_type(&self) -> ExtensionType

Returns the ExtensionType