Struct openmls::extensions::Extensions

pub struct Extensions { /* private fields */ }

A list of extensions with unique extension types.

Implementations

impl Extensions

pub fn empty() -> Self

Create an empty extension list.

pub fn single(extension: Extension) -> Self

Create an extension list with a single extension.

pub fn from_vec( extensions: Vec<Extension>, ) -> Result<Self, InvalidExtensionError>

Create an extension list with multiple extensions.

This function will fail when the list of extensions contains duplicate extension types.

pub fn iter(&self) -> impl Iterator<Item = &Extension>

Returns an iterator over the extension list.

pub fn add(&mut self, extension: Extension) -> Result<(), InvalidExtensionError>

Add an extension to the extension list.

Returns an error when there already is an extension with the same extension type.

pub fn add_or_replace(&mut self, extension: Extension) -> Option<Extension>

Add an extension to the extension list (or replace an existing one.)

Returns the replaced extension (if any).

pub fn remove(&mut self, extension_type: ExtensionType) -> Option<Extension>

Remove an extension from the extension list.

Returns the removed extension or None when there is no extension with the given extension type.

pub fn contains(&self, extension_type: ExtensionType) -> bool

Returns true iff the extension list contains an extension with the given extension type.

impl Extensions

pub fn application_id(&self) -> Option<&ApplicationIdExtension>

Get a reference to the ApplicationIdExtension if there is any.

pub fn ratchet_tree(&self) -> Option<&RatchetTreeExtension>

Get a reference to the RatchetTreeExtension if there is any.

pub fn required_capabilities(&self) -> Option<&RequiredCapabilitiesExtension>

Get a reference to the RequiredCapabilitiesExtension if there is any.

pub fn external_pub(&self) -> Option<&ExternalPubExtension>

Get a reference to the ExternalPubExtension if there is any.

pub fn external_senders(&self) -> Option<&ExternalSendersExtension>

Get a reference to the ExternalSendersExtension if there is any.