Trait openmls::extensions::Extension

pub trait Extension: Debug + ExtensionHelper + Send + Serialize + Deserialize {
    fn new_from_bytes(bytes: &[u8]) -> Result<Self, ExtensionError>
    where
        Self: Sized;

    fn extension_type(&self) -> ExtensionType;

    fn to_extension_struct(&self) -> ExtensionStruct;

    fn as_any(&self) -> &dyn Any;

    fn to_parent_hash_extension(
        &self
    ) -> Result<&ParentHashExtension, ExtensionError> { ... }

    fn to_capabilities_extension(
        &self
    ) -> Result<&CapabilitiesExtension, ExtensionError> { ... }

    fn to_lifetime_extension(
        &self
    ) -> Result<&LifetimeExtension, ExtensionError> { ... }

    fn to_key_id_extension(&self) -> Result<&KeyIDExtension, ExtensionError> { ... }

    fn as_ratchet_tree_extension(
        &self
    ) -> Result<&RatchetTreeExtension, ExtensionError> { ... }
}

Extension

This trait defines functions to interact with an extension.

Required methods

fn new_from_bytes(bytes: &[u8]) -> Result<Self, ExtensionError> where
Self: Sized,

Build a new extension from a byte slice.

Note that all implementations of this trait are not public such that this function can't be used outside of the library.

fn extension_type(&self) -> ExtensionType

Each extension has an extension type. This should be an associated constant really. See https://github.com/rust-lang/rust/issues/46969 for reference.

fn to_extension_struct(&self) -> ExtensionStruct

Get the extension as ExtensionStruct for encoding.

fn as_any(&self) -> &dyn Any

Get a generic trait object for downcasting.

Loading content...
Provided methods

fn to_parent_hash_extension(
&self
) -> Result<&ParentHashExtension, ExtensionError>

Get a reference to the ParentHashExtension. Returns an InvalidExtensionType error if called on an Extension that's not a ParentHashExtension.

fn to_capabilities_extension(
&self
) -> Result<&CapabilitiesExtension, ExtensionError>

Get a reference to the CapabilitiesExtension. Returns an InvalidExtensionType error if called on an Extension that's not a CapabilitiesExtension.

fn to_lifetime_extension(&self) -> Result<&LifetimeExtension, ExtensionError>

Get a reference to the LifetimeExtension. Returns an InvalidExtensionType error if called on an Extension that's not a LifetimeExtension.

fn to_key_id_extension(&self) -> Result<&KeyIDExtension, ExtensionError>

Get a reference to the KeyIDExtension. Returns an InvalidExtensionType error if called on an Extension that's not a KeyIDExtension.

fn as_ratchet_tree_extension(
&self
) -> Result<&RatchetTreeExtension, ExtensionError>

Get a reference to the RatchetTreeExtension. Returns an InvalidExtensionType error if called on an Extension that's not a RatchetTreeExtension.

Loading content...