openmls::framing

Enum ProtocolMessage

pub enum ProtocolMessage {
    PrivateMessage(PrivateMessageIn),
    PublicMessage(Box<PublicMessageIn>),
}

Available on 32-bit or 64-bit only.

Enum containing a message for use with process_message and an MlsGroup. Both PublicMessage and PrivateMessage implement Into<ProtocolMessage>.

Variants

PrivateMessage(PrivateMessageIn)

A ProtocolMessage containing a PrivateMessage.

PublicMessage(Box<PublicMessageIn>)

A ProtocolMessage containing a PublicMessage.

Implementations

impl ProtocolMessage

pub fn wire_format(&self) -> WireFormat

Returns the wire format.

pub fn group_id(&self) -> &GroupId

Returns the group ID.

pub fn epoch(&self) -> GroupEpoch

Returns the epoch.

pub fn content_type(&self) -> ContentType

Returns the content type.

pub fn is_external(&self) -> bool

Returns true if this is either an external proposal or external commit

pub fn is_handshake_message(&self) -> bool

Returns true if this is a handshake message and false otherwise.