Struct openmls::framing::plaintext::MLSPlaintext

pub struct MLSPlaintext {
    pub(crate) group_id: GroupId,
    pub(crate) epoch: GroupEpoch,
    pub(crate) sender: Sender,
    pub(crate) authenticated_data: Vec<u8>,
    pub(crate) content_type: ContentType,
    pub(crate) content: MLSPlaintextContentType,
    pub(crate) signature: Signature,
    pub(crate) confirmation_tag: Option<ConfirmationTag>,
    pub(crate) membership_tag: Option<MembershipTag>,
}

MLSPlaintext is a framing structure for MLS messages. It can contain Proposals, Commits and application messages.

Message framing

struct {
    opaque group_id<0..255>;
    uint64 epoch;
    Sender sender;
    opaque authenticated_data<0..2^32-1>;

    ContentType content_type;
    select (MLSPlaintext.content_type) {
        case application:
            opaque application_data<0..2^32-1>;

        case proposal:
            Proposal proposal;

        case commit:
            Commit commit;
    }

opaque signature<0..2^16-1>;
optional<MAC> confirmation_tag;
optional<MAC> membership_tag;

} MLSPlaintext;

Fields
group_id: GroupIdepoch: GroupEpochsender: Senderauthenticated_data: Vec<u8>content_type: ContentTypecontent: MLSPlaintextContentTypesignature: Signatureconfirmation_tag: Option<ConfirmationTag>membership_tag: Option<MembershipTag>
Implementations

impl MLSPlaintext

pub(crate) fn new_from_member(
sender_index: LeafIndex,
authenticated_data: &[u8],
content: MLSPlaintextContentType,
credential_bundle: &CredentialBundle,
context: &GroupContext
) -> Result<Self, CodecError>

This constructor builds a new MLSPlaintext from the parameters provided. It is only used internally.

pub fn new_from_proposal_member(
ciphersuite: &Ciphersuite,
sender_index: LeafIndex,
authenticated_data: &[u8],
proposal: Proposal,
credential_bundle: &CredentialBundle,
context: &GroupContext,
membership_key: &MembershipKey
) -> Result<Self, CodecError>

This constructor builds an MLSPlaintext containing a Proposal. The sender type is always SenderType::Member.

pub fn new_from_application(
ciphersuite: &Ciphersuite,
sender_index: LeafIndex,
authenticated_data: &[u8],
application_message: &[u8],
credential_bundle: &CredentialBundle,
context: &GroupContext,
membership_key: &MembershipKey
) -> Result<Self, CodecError>

This constructor builds an MLSPlaintext containing an application message. The sender type is always SenderType::Member.

pub fn content(&self) -> &MLSPlaintextContentType

Returns a reference to the content field.

pub fn sender(&self) -> LeafIndex

Get the sender leaf index of this message.

pub fn sign_from_external(&mut self, credential_bundle: &CredentialBundle)

Sign this MLSPlaintext. This populates the signature field. The signature is produced from the private key conatined in the credential bundle.

This should be used when signing messages from external parties.

pub fn sign_from_member(
&mut self,
credential_bundle: &CredentialBundle,
serialized_context: &[u8]
) -> Result<(), CodecError>

Sign this MLSPlaintext and add a membership tag. This populates the signature and membership_tag fields. The signature is produced from the private key conatined in the credential bundle, and the membership_tag is produced using the the membership secret.

This should be used to sign messages from group members.

pub fn add_membership_tag(
&mut self,
ciphersuite: &Ciphersuite,
serialized_context: &[u8],
membership_key: &MembershipKey
) -> Result<(), CodecError>

Adds a membership tag to this MLSPlaintext. The membership_tag is produced using the the membership secret.

This should be used after signing messages from group members.

pub fn verify_signature(
&self,
serialized_context: &[u8],
credential: &Credential
) -> Result<(), VerificationError>

Verify the signature of an MLSPlaintext sent from an external party. Returns Ok(()) if successful or VerificationError otherwise.

pub fn verify_membership_tag(
&self,
ciphersuite: &Ciphersuite,
serialized_context: &[u8],
membership_key: &MembershipKey
) -> Result<(), VerificationError>

Verify the membership tag of an MLSPlaintext sent from member. Returns Ok(()) if successful or VerificationError otherwise.

pub fn verify_from_member(
&self,
ciphersuite: &Ciphersuite,
serialized_context: &[u8],
credential: &Credential,
membership_key: &MembershipKey
) -> Result<(), VerificationError>

Verify the signature and the membership tag of an MLSPlaintext sent from a group member. Returns Ok(()) if successful or VerificationError otherwise.

pub fn as_application_message(&self) -> Result<&[u8], MLSPlaintextError>

Tries to extract an application messages from an MLSPlaintext. Returns MLSPlaintextError::NotAnApplicationMessage if the MLSPlaintext contained something other than an application message.

pub fn is_handshake_message(&self) -> bool

Returns true if this is a handshake message and false otherwise.

pub(crate) fn is_proposal(&self) -> bool

Returns true if this is a proposal message and false otherwise.

pub fn group_id(&self) -> &GroupId

Get the group ID.

pub fn epoch(&self) -> &GroupEpoch

Get the group ID.