Struct openmls::framing::plaintext::MlsPlaintext

pub struct MlsPlaintext { /* fields omitted */ }

MLSPlaintext is a framing structure for MLS messages. It can contain Proposals, Commits and application messages.

Message framing

struct {
    opaque group_id<0..255>;
    uint64 epoch;
    Sender sender;
    opaque authenticated_data<0..2^32-1>;

    ContentType content_type;
    select (MLSPlaintext.content_type) {
        case application:
            opaque application_data<0..2^32-1>;

        case proposal:
            Proposal proposal;

        case commit:
            Commit commit;
    }

    opaque signature<0..2^16-1>;
    optional<MAC> confirmation_tag;
    optional<MAC> membership_tag;
} MLSPlaintext;

Implementations

impl MlsPlaintext

pub fn new_proposal(
sender_index: LeafIndex,
authenticated_data: &[u8],
proposal: Proposal,
credential_bundle: &CredentialBundle,
context: &GroupContext,
membership_key: &MembershipKey
) -> Result<Self, MlsPlaintextError>

This constructor builds an MlsPlaintext containing a Proposal. The sender type is always SenderType::Member.

pub fn new_commit(
sender_index: LeafIndex,
authenticated_data: &[u8],
commit: Commit,
credential_bundle: &CredentialBundle,
context: &GroupContext
) -> Result<Self, MlsPlaintextError>

This constructor builds an MlsPlaintext containing a Commit. The sender type is always SenderType::Member.

pub fn new_application(
sender_index: LeafIndex,
authenticated_data: &[u8],
application_message: &[u8],
credential_bundle: &CredentialBundle,
context: &GroupContext,
membership_key: &MembershipKey
) -> Result<Self, MlsPlaintextError>

This constructor builds an MlsPlaintext containing an application message. The sender type is always SenderType::Member.

pub fn content(&self) -> &MlsPlaintextContentType

Returns a reference to the content field.

pub fn sender(&self) -> &Sender

Get the sender of this message.

pub fn sender_index(&self) -> LeafIndex

Get the sender leaf index of this message.

pub fn verify_membership(
&self,
serialized_context: &[u8],
membership_key: &MembershipKey
) -> Result<(), MlsPlaintextError>

Verify the membership tag of an MlsPlaintext sent from a group member. Returns Ok(()) if successful or VerificationError otherwise. Note that the signature must have been verified before.

pub fn as_application_message(&self) -> Result<&[u8], MlsPlaintextError>

Tries to extract an application messages from an MlsPlaintext. Returns MlsPlaintextError::NotAnApplicationMessage if the MlsPlaintext contained something other than an application message.

pub fn is_handshake_message(&self) -> bool

Returns true if this is a handshake message and false otherwise.

pub fn group_id(&self) -> &GroupId

Get the group ID.

pub fn epoch(&self) -> &GroupEpoch

Get the group ID.

pub fn authenticated_data(&self) -> &[u8]

The the authenticated data of this MlsPlaintext as byte slice.