Module openmls::group::config

Configuration module for MlsGroup configurations.

Building an MlsGroupConfig

The MlsGroupConfigBuilder makes it easy to build configurations for the MlsGroup.

use openmls::prelude::*;

let group_config = MlsGroupConfig::builder()
    .use_ratchet_tree_extension(true)
    .build();

See MlsGroupConfigBuilder for all options that can be configured.

Wire format policies

Only some combination of possible wire formats are valid within OpenMLS. The WIRE_FORMAT_POLICIES lists all valid options that can be set.

use openmls::prelude::*;

let group_config = MlsGroupConfig::builder()
    .wire_format_policy(MIXED_CIPHERTEXT_WIRE_FORMAT_POLICY)
    .build();

Structs

MlsGroupConfig

Specifies the configuration parameters for a MlsGroup. Refer to the User Manual for more information about the different configuration values.

MlsGroupConfigBuilder

Builder for an MlsGroupConfig.

WireFormatPolicy

Defines what wire format is desired for outgoing handshake messages. Note that application messages must always be encrypted.

Enums

IncomingWireFormatPolicy

Defines what wire format is acceptable for incoming handshake messages. Note that application messages must always be encrypted.

OutgoingWireFormatPolicy

Defines what wire format should be used for outgoing handshake messages. Note that application messages must always be encrypted.

Constants

MIXED_CIPHERTEXT_WIRE_FORMAT_POLICY

Incoming wire formats can be mixed while outgoing wire formats are always ciphertext.

MIXED_PLAINTEXT_WIRE_FORMAT_POLICY

Incoming wire formats can be mixed while outgoing wire formats are always plaintext.

PURE_CIPHERTEXT_WIRE_FORMAT_POLICY

Incoming and outgoing wire formats are always ciphertext.

PURE_PLAINTEXT_WIRE_FORMAT_POLICY

Incoming and outgoing wire formats are always plaintext.

WIRE_FORMAT_POLICIES

All valid wire format policy combinations.