openmls::group::dmls::dmls_group

Struct DmlsGroup

pub struct DmlsGroup(/* private fields */);

The DmlsGroup struct is a wrapper around MlsGroup that provides DMLs-specific functionality.

Implementations

impl DmlsGroup

pub fn new<Provider: OpenDmlsProvider>( provider: &Provider, signer: &impl Signer, mls_group_create_config: &MlsGroupCreateConfig, credential_with_key: CredentialWithKey, ) -> Result<Self, NewGroupError<<Provider as OpenMlsProvider>::StorageError>>

Creates a new DmlsGroup instance.

pub fn from_staged_welcome<Provider: OpenDmlsProvider>( provider: &Provider, staged_welcome: StagedWelcome, ) -> Result<Self, WelcomeError<<Provider as OpenMlsProvider>::StorageError>>

Creates a new DmlsGroup instance from a StagedWelcome.

pub fn merge_staged_commit<Provider: OpenDmlsProvider>( &mut self, provider: &Provider, staged_commit: StagedCommit, ) -> Result<(), DmlsMergeError<<Provider as OpenMlsProvider>::StorageError>>

Merge a StagedCommit into the group after inspection. As this advances the epoch of the group, it also clears any pending commits.

pub fn merge_pending_commit<Provider: OpenDmlsProvider>( &mut self, provider: &Provider, ) -> Result<(), DmlsMergePendingError<<Provider as OpenMlsProvider>::StorageError>>

Merge a pending commit into the group.

pub fn derive_epoch_id<Provider: OpenMlsCrypto>( &self, provider: &Provider, ) -> Result<DmlsEpoch, ExportSecretError>

Derive the current epoch id of this group.

pub fn load_for_epoch<Provider: DmlsStorageProvider>( storage: &Provider, epoch: DmlsEpoch, group_id: &GroupId, ) -> Option<Self>

Returns the group state of the group with the given group id for the given epoch.

impl DmlsGroup

pub fn add_members<Provider: OpenDmlsProvider>( &mut self, provider: &Provider, signer: &impl Signer, key_packages: &[KeyPackage], ) -> Result<(DmlsMessageOut, MlsMessageOut, Option<GroupInfo>), AddMembersError<<Provider as OpenMlsProvider>::StorageError>>

Add members to the group.

pub fn process_message<Provider: OpenDmlsProvider>( &mut self, provider: &Provider, message: DmlsMessageIn, ) -> Result<ProcessedMessage, ProcessDmlsMessageError<<Provider as OpenMlsProvider>::StorageError>>

Load the correct group and process a message.

pub fn self_update<Provider: OpenDmlsProvider>( &mut self, provider: &Provider, signer: &impl Signer, leaf_node_parameters: LeafNodeParameters, ) -> Result<DmlsCommitMessageBundle, SelfUpdateError<<Provider as OpenMlsProvider>::StorageError>>

DMLS wrapper around the MlsGroup::self_update function.

pub fn clear_pending_commit<Provider: OpenDmlsProvider>( &mut self, provider: &Provider, ) -> Result<(), <Provider as OpenMlsProvider>::StorageError>

Clear a pending commit