Enum openmls::group::errors::ApplyCommitError

pub enum ApplyCommitError {
    PlaintextSignatureFailure(VerificationError),
    DecryptionFailure(TreeError),
    CodecError(CodecError),
    KeyScheduleError(KeyScheduleError),
    PskError(PskError),
    EpochMismatch,
    WrongPlaintextContentType,
    SelfRemoved,
    PathKeyPackageVerificationFailure,
    NoParentHashExtension,
    ParentHashMismatch,
    RequiredPathNotFound,
    ConfirmationTagMissing,
    ConfirmationTagMismatch,
    MissingOwnKeyPackage,
    MissingProposal,
    OwnKeyNotFound,
    InitSecretNotFound,
}

Variants

PlaintextSignatureFailure(VerificationError)

MlsPlaintext signature is invalid.

DecryptionFailure(TreeError)

A matching EncryptedPathSecret failed to decrypt.

CodecError(CodecError)

Codec error occurred.

KeyScheduleError(KeyScheduleError)

An error occurred in the key schedule.

PskError(PskError)

A PSK error occured.

EpochMismatch

Couldn't apply Commit. The epoch of the group context and MlsPlaintext didn't match.

WrongPlaintextContentType

apply_commit_internal was called with an MlsPlaintext that is not a Commit.

SelfRemoved

Tried to apply a commit to a group we are not a part of.

PathKeyPackageVerificationFailure

Unable to verify the key package signature.

NoParentHashExtension

Parent hash extension is missing.

ParentHashMismatch

Parent hash values don't match.

RequiredPathNotFound

Unable to determine commit path.

ConfirmationTagMissing

Confirmation Tag is missing.

ConfirmationTagMismatch

Confirmation tag is invalid.

MissingOwnKeyPackage

No key package provided to apply own commit.

MissingProposal

The proposal queue is missing a proposal for the commit.

OwnKeyNotFound

Missing own key to apply proposal.

InitSecretNotFound

Missing init secret to apply proposal.

Implementations

impl ApplyCommitError

pub(crate) fn _description(&self) -> String