Enum openmls::group::errors::CoreGroupError

pub enum CoreGroupError {

    MlsCiphertextError(MlsCiphertextError),
    MlsPlaintextError(MlsPlaintextError),
    WelcomeError(WelcomeError),
    ExternalCommitError(ExternalCommitError),
    StageCommitError(StageCommitError),
    CreateCommitError(CreateCommitError),
    ConfigError(ConfigError),
    ExporterError(ExporterError),
    ProposalQueueError(ProposalQueueError),
    CodecError(TlsCodecError),
    KeyScheduleError(KeyScheduleError),
    PskSecretError(PskSecretError),
    CredentialError(CredentialError),
    TreeError(TreeError),
    TreeSyncError(TreeSyncError),
    TreeSyncDiffError(TreeSyncDiffError),
    TreeKemError(TreeKemError),
    KeyPackageError(KeyPackageError),
    ExtensionError(ExtensionError),
    ValidationError(ValidationError),
    FramingValidationError(FramingValidationError),
    ProposalValidationError(ProposalValidationError),
    ExternalCommitValidationError(ExternalCommitValidationError),
    CryptoError(CryptoError),
    InterimTranscriptHashError(InterimTranscriptHashError),
    QueuedProposalError(QueuedProposalError),
    MissingKeyPackageBundle,
    NoSignatureKey,
    OwnCommitError,
    LibraryError,

}

Variants

MlsCiphertextError(MlsCiphertextError)

Tuple Fields

0: MlsCiphertextError

See MlsCiphertextError for details.

MlsPlaintextError(MlsPlaintextError)

Tuple Fields

0: MlsPlaintextError

See MlsPlaintextError for details.

WelcomeError(WelcomeError)

Tuple Fields

0: WelcomeError

See WelcomeError for details.

ExternalCommitError(ExternalCommitError)

Tuple Fields

0: ExternalCommitError

See Externaallow(lint) for details.

StageCommitError(StageCommitError)

Tuple Fields

0: StageCommitError

See StageCommitError for details.

CreateCommitError(CreateCommitError)

Tuple Fields

0: CreateCommitError

See CreateCommitError for details.

ConfigError(ConfigError)

Tuple Fields

0: ConfigError

See ConfigError for details.

ExporterError(ExporterError)

Tuple Fields

0: ExporterError

See ExporterError for details.

ProposalQueueError(ProposalQueueError)

Tuple Fields

0: ProposalQueueError

See ProposalQueueError for details.

CodecError(TlsCodecError)

Tuple Fields

0: TlsCodecError

TLS (de)serialization error occurred.

KeyScheduleError(KeyScheduleError)

Tuple Fields

0: KeyScheduleError

An error occurred in the key schedule.

PskSecretError(PskSecretError)

Tuple Fields

0: PskSecretError

A PskSecret error occurred.

CredentialError(CredentialError)

Tuple Fields

0: CredentialError

See CredentialError for details.

TreeError(TreeError)

Tuple Fields

0: TreeError

See TreeError for details.

TreeSyncError(TreeSyncError)

Tuple Fields

0: TreeSyncError

See TreeSyncError for details.

TreeSyncDiffError(TreeSyncDiffError)

Tuple Fields

0: TreeSyncDiffError

See TreeSyncDiffError for details.

TreeKemError(TreeKemError)

Tuple Fields

0: TreeKemError

See TreeKemError for details.

KeyPackageError(KeyPackageError)

Tuple Fields

0: KeyPackageError

See KeyPackageError for details.

ExtensionError(ExtensionError)

Tuple Fields

0: ExtensionError

See ExtensionError for details.

ValidationError(ValidationError)

Tuple Fields

0: ValidationError

See ValidationError for details.

FramingValidationError(FramingValidationError)

Tuple Fields

0: FramingValidationError

See FramingValidationError for details.

ProposalValidationError(ProposalValidationError)

Tuple Fields

0: ProposalValidationError

See ProposalValidationError for details.

ExternalCommitValidationError(ExternalCommitValidationError)

Tuple Fields

0: ExternalCommitValidationError

See ProposalValidationError for details.

CryptoError(CryptoError)

Tuple Fields

0: CryptoError

See CryptoError for details.

InterimTranscriptHashError(InterimTranscriptHashError)

Tuple Fields

0: InterimTranscriptHashError

See InterimTranscriptHashError for details.

QueuedProposalError(QueuedProposalError)

Tuple Fields

0: QueuedProposalError

See QueuedProposalError for details.

MissingKeyPackageBundle

Couldn’t find KeyPackageBundle corresponding to own update proposal.

NoSignatureKey

No signature key was found.

OwnCommitError

Can’t process a commit created by the owner of the group. Please merge the [StagedCommit] returned by create_commit instead.

LibraryError

An unrecoverable error has occurred due to a bug in the implementation.