Enum openmls::group::errors::WelcomeError

pub enum WelcomeError<KeyStoreError> {

    LibraryError(LibraryError),
    CiphersuiteMismatch,
    GroupInfoCiphersuiteMismatch,
    JoinerSecretNotFound,
    MissingRatchetTree,
    ConfirmationTagMismatch,
    InvalidGroupInfoSignature,
    GroupInfoDecryptionFailure,
    UnsupportedMlsVersion,
    UnsupportedCapability,
    UnknownSender,
    MalformedWelcomeMessage,
    UnableToDecrypt,
    UnsupportedExtensions,
    PskTooManyKeys,
    PskNotFound,
    NoMatchingEncryptionKey,
    NoMatchingKeyPackage,
    KeyStoreError(KeyStoreError),
    PublicTreeError(PublicTreeError),
    LeafNodeValidation(LeafNodeValidationError),

}

Welcome error

Variants

LibraryError(LibraryError)

See LibraryError for more details.

CiphersuiteMismatch

Ciphersuites in Welcome and key package bundle don’t match.

GroupInfoCiphersuiteMismatch

Ciphersuites in Welcome/GroupInfo and key package bundle don’t match.

JoinerSecretNotFound

No joiner secret found in the Welcome message.

MissingRatchetTree

No ratchet tree available to build initial tree after receiving a Welcome message.

ConfirmationTagMismatch

The computed confirmation tag does not match the expected one.

InvalidGroupInfoSignature

The signature on the GroupInfo is not valid.

GroupInfoDecryptionFailure

Unable to decrypt the GroupInfo.

UnsupportedMlsVersion

We don’t support the version of the group we are trying to join.

UnsupportedCapability

We don’t support all capabilities of the group.

UnknownSender

Sender not found in tree.

MalformedWelcomeMessage

Malformed Welcome message.

UnableToDecrypt

Could not decrypt the Welcome message.

UnsupportedExtensions

Unsupported extensions found in the KeyPackage of another member.

PskTooManyKeys

More than 2^16 PSKs were provided.

PskNotFound

The PSK could not be found in the key store.

NoMatchingEncryptionKey

No matching encryption key was found in the key store.

NoMatchingKeyPackage

No matching key package was found in the key store.

KeyStoreError(KeyStoreError)

Error accessing the key store.

PublicTreeError(PublicTreeError)

This error indicates the public tree is invalid. See PublicTreeError for more details.

LeafNodeValidation(LeafNodeValidationError)

This error indicates the leaf node is invalid. See LeafNodeValidationError for more details.