Module openmls::group::events

Structs

ApplicationMessageEvent
Event that occurs when an application message is received. sender contains the message’s sender and message contains the application message.

ErrorEvent
Event that occurs when an error occurred while processing messages in a group. error contains the specific error that occurred.

MemberAddedEvent
Event that occurs when member sender adds member added_member.

MemberRemovedEvent
Event that occurs when a member leaves/is removed from the group. The exact context of the removal is explained in the Removal enum.

MemberUpdatedEvent
Event that occurs when a member updates its leaf. updated_member contains the new credential.

PskReceivedEvent
Event that occurs when a PSK is received. psk_id contains the PSK ID.

ReInitEvent
Event that occurs when a ReInitProposal is received. re_init_proposal contains the ReInitProposal.

Enums

GroupEvent
Group event that occured while processing messages in ManagedGroup.

Removal
This enum lists the 4 different variants of a removal, depending on who the remover and who the leaver is.