Module openmls::group::mls_group

MLS Group

This module contains MlsGroup and its submodules.

Modules

config

Configuration module for MlsGroup configurations.

errors

MlsGroup errors

membership

MLS group membership

processing

Processing functions of an MlsGroup for incoming messages.

Structs

MlsGroup

A MlsGroup represents an MLS group with a high-level API. The API exposes high level functions to manage a group by adding/removing members, get the current member list, etc.

Enums

InnerState

Enum that indicates whether the inner group state has been modified since the last time it was persisted. InnerState::Changed indicates that the state has changed and that [.save()] should be called. InnerState::Persisted indicates that the state has not been modified and therefore doesn’t need to be persisted.

MlsGroupState

MlsGroupState determines the state of an MlsGroup. The different states and their transitions are as follows:

PendingCommitState

Pending Commit state. Differentiates between Commits issued by group members and External Commits.