Module openmls::group::mls_group::errors

MlsGroup errors

This module defines the public errors that can be returned from all calls to methods of MlsGroup.

Enums

AddMembersError

Add members error

CommitToPendingProposalsError

Commit to pending proposals error

CreateMessageError

Create message error

EmptyInputError

EmptyInput error

ExportPublicGroupStateError

Export public group state error

ExportSecretError

Export secret error

LeaveGroupError

Leave group error

MlsGroupStateError

Group state error

NewGroupError

New group error

ParseMessageError

Parse message error

ProposeAddMemberError

Propose add members error

ProposeRemoveMemberError

Propose remove members error

ProposeSelfUpdateError

Propose self update error

RemoveMembersError

Remove members error

SelfUpdateError

Self update error

UnverifiedMessageError

Unverified message error