Module openmls::group::mls_group::proposals

Structs

CreationProposalQueue

Proposal queue that helps filtering and sorting the Proposals from one epoch. The Proposals are stored in a HashMap which maps Proposal references to Proposals, such that, given a reference, a proposal can be accessed efficiently. To enable iteration over the queue in order, the ProposalQueue also contains a vector of ProposalReferences.

ProposalStore

A ProposalStore can store the standalone proposals that are received from the DS in between two commit messages.

StagedAddProposal

A staged Add proposal

StagedProposal

Alternative representation of a Proposal, where the sender is extracted from the encapsulating MlsPlaintext and the ProposalReference is attached.

StagedProposalQueue

Proposal queue that helps filtering and sorting the staged Proposals from one epoch. The Proposals are stored in a HashMap which maps Proposal references to Proposals, such that, given a reference, a proposal can be accessed efficiently. To enable iteration over the queue in order, the ProposalQueue also contains a vector of ProposalReferences.

StagedPskProposal

A staged PresharedKey proposal

StagedRemoveProposal

A staged Remove proposal

StagedUpdateProposal

A staged Update proposal