Struct openmls::group::mls_group::MlsGroup

pub struct MlsGroup { /* fields omitted */ }

Implementations

impl MlsGroup

pub fn create_commit(
&self,
params: CreateCommitParams<'_>,
backend: &impl OpenMlsCryptoProvider
) -> CreateCommitResult

impl MlsGroup

pub fn stage_commit(
&mut self,
mls_plaintext: &MlsPlaintext,
proposal_store: &ProposalStore,
own_key_packages: &[KeyPackageBundle],
psk_fetcher_option: Option<PskFetcher>,
backend: &impl OpenMlsCryptoProvider
) -> Result<StagedCommit, MlsGroupError>

Stages a commit message. This function does the following:

Applies the proposals covered by the commit to the tree

Applies the (optional) update path to the tree

Calculates the path secrets

Initializes the key schedule for epoch rollover

Verifies the confirmation tag/membership tag Returns a StagedCommit that can be inspected and later merged into the group state with [merge_commit()]

pub fn merge_commit(&mut self, staged_commit: StagedCommit)

Merges a StagedCommit into the group state.

pub fn cancel_commit(&mut self, staged_commit: StagedCommit)

This is temporary and will disappear when #424 is addressed. This is just here for completeness but won’t be used anywhere. Rolls back the public tree nodes in case a Commit contained undesired proposals.

impl MlsGroup

Public MlsGroup functions.

pub fn new(
id: &[u8],
ciphersuite_name: CiphersuiteName,
backend: &impl OpenMlsCryptoProvider,
key_package_bundle: KeyPackageBundle,
config: MlsGroupConfig,
psk_option: impl Into<Option<PskSecret>>,
version: impl Into<Option<ProtocolVersion>>
) -> Result<Self, MlsGroupError>

pub fn new_from_welcome(
welcome: Welcome,
nodes_option: Option<Vec<Option<Node>>>,
kpb: KeyPackageBundle,
psk_fetcher_option: Option<PskFetcher>,
backend: &impl OpenMlsCryptoProvider
) -> Result<Self, MlsGroupError>

pub fn create_add_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
joiner_key_package: KeyPackage,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

pub fn create_update_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
key_package: KeyPackage,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

pub fn create_remove_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
removed_index: LeafIndex,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

pub fn create_presharedkey_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
psk: PreSharedKeyId,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

pub fn create_application_message(
&mut self,
aad: &[u8],
msg: &[u8],
credential_bundle: &CredentialBundle,
padding_size: usize,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsCiphertext, MlsGroupError>

pub fn encrypt(
&mut self,
mls_plaintext: MlsPlaintext,
padding_size: usize,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsCiphertext, MlsGroupError>

pub fn decrypt(
&mut self,
mls_ciphertext: &MlsCiphertext,
backend: &impl OpenMlsCryptoProvider
) -> Result<VerifiableMlsPlaintext, MlsGroupError>

Decrypt an MlsCiphertext into an MlsPlaintext

pub fn verify(
&self,
verifiable: VerifiableMlsPlaintext,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

Set the context of the VerifiableMlsPlaintext (if it has not been set already), verify it and return the MlsPlaintext.

pub fn verify_membership_tag(
&self,
backend: &impl OpenMlsCryptoProvider,
verifiable_mls_plaintext: &mut VerifiableMlsPlaintext
) -> Result<(), MlsGroupError>

Set the context of the UnverifiedMlsPlaintext and verify its membership tag.

pub fn export_secret(
&self,
backend: &impl OpenMlsCryptoProvider,
label: &str,
context: &[u8],
key_length: usize
) -> Result<Vec<u8>, MlsGroupError>

Exporter

pub fn authentication_secret(&self) -> Vec<u8>

Returns the authentication secret

pub fn load<R: Read>(reader: R) -> Result<MlsGroup, Error>

Loads the state from persisted state

pub fn save<W: Write>(&self, writer: &mut W) -> Result<(), Error>

Persists the state

pub fn tree(&self) -> Ref<'_, RatchetTree>

Returns the ratchet tree

pub fn ciphersuite(&self) -> &'static Ciphersuite

Get the ciphersuite implementation used in this group.

pub fn context(&self) -> &GroupContext

Get the group context

pub fn group_id(&self) -> &GroupId

Get the group ID

pub fn extensions(&self) -> Vec<Extension>

Get the groups extensions. Right now this is limited to the ratchet tree extension which is built on the fly when calling this function.

pub fn export_public_group_state(
&self,
backend: &impl OpenMlsCryptoProvider,
credential_bundle: &CredentialBundle
) -> Result<PublicGroupState, CredentialError>

Export the PublicGroupState

pub fn use_ratchet_tree_extension(&self) -> bool

Returns true if the group uses the ratchet tree extension anf `false otherwise