Struct openmls::group::CoreGroup

pub struct CoreGroup { /* fields omitted */ }

Implementations

impl CoreGroup

pub fn validate_framing(
&self,
message: &MlsMessageIn
) -> Result<(), CoreGroupError>

Checks the following semantic validation:

ValSem2

ValSem3

pub fn validate_plaintext(
&self,
plaintext: &VerifiableMlsPlaintext
) -> Result<(), CoreGroupError>

Checks the following semantic validation:

ValSem4

ValSem5

ValSem7

ValSem9

pub fn validate_add_proposals(
&self,
staged_proposal_queue: &StagedProposalQueue
) -> Result<(), CoreGroupError>

Validate Add proposals. This function implements the following checks:

ValSem100

ValSem101

ValSem102

ValSem103

ValSem104

ValSem105

TODO: ValSem106

pub fn validate_remove_proposals(
&self,
staged_proposal_queue: &StagedProposalQueue
) -> Result<(), CoreGroupError>

Validate Remove proposals. This function implements the following checks:

ValSem107

ValSem108

pub fn validate_update_proposals(
&self,
staged_proposal_queue: &StagedProposalQueue,
path_key_package: Option<(LeafIndex, &KeyPackage)>
) -> Result<(), CoreGroupError>

Validate Update proposals. This function implements the following checks:

ValSem109

ValSem110

impl CoreGroup

pub fn create_commit(
&self,
params: CreateCommitParams<'_>,
backend: &impl OpenMlsCryptoProvider
) -> CreateCommitResult

impl CoreGroup

pub fn parse_message<'a>(
&mut self,
message: MlsMessageIn,
message_secrets_store: impl Into<Option<&'a mut MessageSecretsStore>>,
backend: &impl OpenMlsCryptoProvider
) -> Result<UnverifiedMessage, CoreGroupError>

This function is used to parse messages from the DS. It checks for syntactic errors and makes some semantic checks as well. If the input is a MlsCiphertext message, it will be decrypted. Returns an UnverifiedMessage that can be inspected and later processed in Self::process_unverified_message(). Checks the following semantic validation:

ValSem2

ValSem3

ValSem4

ValSem5

ValSem6

ValSem7

ValSem9

pub fn process_unverified_message<'a>(
&mut self,
unverified_message: UnverifiedMessage,
signature_key: Option<&SignaturePublicKey>,
proposal_store: &ProposalStore,
message_secrets_store: impl Into<Option<&'a mut MessageSecretsStore>>,
own_kpbs: &[KeyPackageBundle],
backend: &impl OpenMlsCryptoProvider
) -> Result<ProcessedMessage, CoreGroupError>

This processing function does most of the semantic verifications. It returns a ProcessedMessage enum. Checks the following semantic validation:

ValSem8

ValSem10

ValSem100

ValSem101

ValSem102

ValSem103

ValSem104

ValSem105

ValSem106

ValSem107

ValSem109

ValSem110

pub fn merge_staged_commit(
&mut self,
staged_commit: StagedCommit,
proposal_store: &mut ProposalStore,
message_secrets_store: &mut MessageSecretsStore
) -> Result<(), CoreGroupError>

Merge a [StagedCommit] into the group after inspection

impl CoreGroup

pub fn stage_commit(
&mut self,
mls_plaintext: &MlsPlaintext,
proposal_store: &ProposalStore,
own_key_packages: &[KeyPackageBundle],
backend: &impl OpenMlsCryptoProvider
) -> Result<StagedCommit, CoreGroupError>

Stages a commit message. This function does the following:

Applies the proposals covered by the commit to the tree

Applies the (optional) update path to the tree

Calculates the path secrets

Initializes the key schedule for epoch rollover

Verifies the confirmation tag/membership tag Returns a [StagedCommit] that can be inspected and later merged into the group state with [CoreGroup::merge_commit()] This function does the following checks:

ValSem100

ValSem101

ValSem102

ValSem103

ValSem104

ValSem105

ValSem106

ValSem107

ValSem108

ValSem109

ValSem110

ValSem201

ValSem205

pub fn merge_commit_take_message_secrets(
&mut self,
staged_commit: StagedCommit
) -> Result<Option<MessageSecrets>, CoreGroupError>

Merges a [StagedCommit] into the group state and optionally return a [SecretTree] from the previous epoch. The secret tree is returned if the Commit does not contain a self removal.

This function should not fail and only returns a Result, because it might throw a LibraryError.

impl CoreGroup

Public CoreGroup functions.

pub fn builder(
group_id: GroupId,
key_package_bundle: KeyPackageBundle
) -> CoreGroupBuilder

Get a builder for CoreGroup.

pub fn new_from_welcome(
welcome: Welcome,
nodes_option: Option<Vec<Option<Node>>>,
kpb: KeyPackageBundle,
backend: &impl OpenMlsCryptoProvider
) -> Result<Self, CoreGroupError>

pub fn create_add_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
joiner_key_package: KeyPackage,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, CoreGroupError>

pub fn create_update_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
key_package: KeyPackage,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, CoreGroupError>

pub fn create_remove_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
removed_index: LeafIndex,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, CoreGroupError>

pub fn create_application_message(
&mut self,
aad: &[u8],
msg: &[u8],
credential_bundle: &CredentialBundle,
padding_size: usize,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsCiphertext, CoreGroupError>

pub fn encrypt(
&mut self,
mls_plaintext: MlsPlaintext,
padding_size: usize,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsCiphertext, CoreGroupError>

pub fn export_secret(
&self,
backend: &impl OpenMlsCryptoProvider,
label: &str,
context: &[u8],
key_length: usize
) -> Result<Vec<u8>, CoreGroupError>

Exporter

pub fn authentication_secret(&self) -> Vec<u8>

Returns the authentication secret

pub fn treesync(&self) -> &TreeSync

Returns the ratchet tree

pub fn ciphersuite(&self) -> &'static Ciphersuite

Get the ciphersuite implementation used in this group.

pub fn context(&self) -> &GroupContext

Get the group context

pub fn group_id(&self) -> &GroupId

Get the group ID

pub fn group_context_extensions(&self) -> &[Extension]

Get the group context extensions.

pub fn required_capabilities(&self) -> Option<&RequiredCapabilitiesExtension>

Get the required capabilities extension of this group.