Struct openmls::group::ManagedGroupCallbacks

pub struct ManagedGroupCallbacks {
    pub(crate) validate_add: Option<ValidateAdd>,
    pub(crate) validate_remove: Option<ValidateRemove>,
    pub(crate) auto_save: Option<AutoSave>,
    pub(crate) member_added: Option<MemberAdded>,
    pub(crate) member_removed: Option<MemberRemoved>,
    pub(crate) member_updated: Option<MemberUpdated>,
    pub(crate) psk_received: Option<PskReceived>,
    pub(crate) reinit_received: Option<ReInitReceived>,
    pub(crate) app_message_received: Option<AppMessageReceived>,
    pub(crate) invalid_message_received: Option<InvalidMessageReceived>,
    pub(crate) error_occurred: Option<ErrorOccurred>,
}

Collection of callback functions that are passed to a ManagedGroup as part of the configurations. All callback functions are optional.

Validators

Validator callback functions are called when Proposals are processed through process_messages(). If no validator function is specified for a certain proposal type, any semantically valid proposal will be accepted. Validator functions must return a bool, that indicates whether the proposal is accepted by the application policy.

true means the proposal should be accepted

false means the proposal should be rejected

The validator functions are:

pub type ValidateAdd =
    fn(managed_group: &ManagedGroup, sender: &Credential, added_member: &Credential) -> bool;
pub type ValidateRemove =
    fn(managed_group: &ManagedGroup, sender: &Credential, removed_member: &Credential) -> bool;

Auto-save

The auto-save callback is called whenever the group state was modified and needs to be persisted. The callback function should then call managed_group.save() to persist the group state.

pub type AutoSave = fn(managed_group: &ManagedGroup);

Event listeners

Event listeners get called when certain messages are parsed, or other events occur. The event listeners are:

// Membership changes
pub type MemberAdded =
    fn(managed_group: &ManagedGroup, aad: &[u8], sender: &Credential, added_member: &Credential);
pub type MemberRemoved = fn(managed_group: &ManagedGroup, aad: &[u8], removal: &Removal);
pub type MemberUpdated = fn(managed_group: &ManagedGroup, aad: &[u8], updated_member: &Credential);
// PSKs
pub type PskReceived = fn(managed_group: &ManagedGroup, aad: &[u8], psk_id: &PreSharedKeyID);
pub type ReInitReceived = fn(managed_group: &ManagedGroup, aad: &[u8], re_init: &ReInitProposal);
// Application messages
pub type AppMessageReceived =
    fn(managed_group: &ManagedGroup, aad: &[u8], sender: &Credential, message: &[u8]);
// Errors
pub type InvalidMessageReceived = fn(managed_group: &ManagedGroup, error: InvalidMessageError);
pub type ErrorOccurred = fn(managed_group: &ManagedGroup, error: ManagedGroupError);

Fields
validate_add: Option<ValidateAdd>validate_remove: Option<ValidateRemove>auto_save: Option<AutoSave>member_added: Option<MemberAdded>member_removed: Option<MemberRemoved>member_updated: Option<MemberUpdated>psk_received: Option<PskReceived>reinit_received: Option<ReInitReceived>app_message_received: Option<AppMessageReceived>invalid_message_received: Option<InvalidMessageReceived>error_occurred: Option<ErrorOccurred>
Implementations

impl<'a> ManagedGroupCallbacks

pub fn new() -> Self

pub fn with_validate_add(self, validate_add: ValidateAdd) -> Self

Validator function for AddProposals

pub fn with_validate_remove(self, validate_remove: ValidateRemove) -> Self

Validator function for RemoveProposals

pub fn with_auto_save(self, auto_save: AutoSave) -> Self

Auto-save callback

pub fn with_member_added(self, member_added: MemberAdded) -> Self

Event listener function for AddProposals

pub fn with_member_removed(self, member_removed: MemberRemoved) -> Self

Event listener function for RemoveProposals when a member was removed

pub fn with_member_updated(self, member_updated: MemberUpdated) -> Self

Event listener function for UpdateProposals

pub fn with_psk_received(self, psk_received: PskReceived) -> Self

Event listener function for PreSharedKeyProposals

pub fn with_reinit_received(self, reinit_received: ReInitReceived) -> Self

Event listener function for ReInitProposals

pub fn with_app_message_received(
self,
app_message_received: AppMessageReceived
) -> Self

Event listener function for application messages

pub fn with_invalid_message_received(
self,
invalid_message_received: InvalidMessageReceived
) -> Self

Event listener function for invalid messages

pub fn with_error_occurred(self, error_occurred: ErrorOccurred) -> Self

Event listener function for errors that occur