Struct openmls::group::MlsGroupConfigBuilder

pub struct MlsGroupConfigBuilder { /* private fields */ }

Builder for an MlsGroupConfig.

Implementations

impl MlsGroupConfigBuilder

pub fn new() -> Self

Creates a new builder with default values.

pub fn wire_format_policy(self, wire_format_policy: WireFormatPolicy) -> Self

Sets the wire_format property of the MlsGroupConfig.

pub fn padding_size(self, padding_size: usize) -> Self

Sets the padding_size property of the MlsGroupConfig.

pub fn max_past_epochs(self, max_past_epochs: usize) -> Self

Sets the max_past_epochs property of the MlsGroupConfig. This allows application messages from previous epochs to be decrypted.

WARNING

This feature enables the storage of message secrets from past epochs. It is a trade-off between functionality and forward secrecy and should only be enabled if the Delivery Service cannot guarantee that application messages will be sent in the same epoch in which they were generated. The number for max_epochs should be as low as possible.

pub fn number_of_resumtion_secrets(
self,
number_of_resumption_secrets: usize
) -> Self

Sets the number_of_resumption_secrets property of the MlsGroupConfig.

pub fn use_ratchet_tree_extension(
self,
use_ratchet_tree_extension: bool
) -> Self

Sets the use_ratchet_tree_extension property of the MlsGroupConfig.

pub fn sender_ratchet_configuration(
self,
sender_ratchet_configuration: SenderRatchetConfiguration
) -> Self

Sets the sender_ratchet_configuration property of the MlsGroupConfig. See SenderRatchetConfiguration for more information.

pub fn build(self) -> MlsGroupConfig

Finalizes the builder and retursn an [MlsGroupConfig].