Struct openmls::group::MlsGroupCreateConfigBuilder

pub struct MlsGroupCreateConfigBuilder { /* private fields */ }

Builder for an MlsGroupCreateConfig.

Implementations

impl MlsGroupCreateConfigBuilder

pub fn wire_format_policy(self, wire_format_policy: WireFormatPolicy) -> Self

Sets the wire_format property of the MlsGroupCreateConfig.

pub fn padding_size(self, padding_size: usize) -> Self

Sets the padding_size property of the MlsGroupCreateConfig.

pub fn max_past_epochs(self, max_past_epochs: usize) -> Self

Sets the max_past_epochs property of the MlsGroupCreateConfig. This allows application messages from previous epochs to be decrypted.

WARNING

This feature enables the storage of message secrets from past epochs. It is a trade-off between functionality and forward secrecy and should only be enabled if the Delivery Service cannot guarantee that application messages will be sent in the same epoch in which they were generated. The number for max_epochs should be as low as possible.

pub fn number_of_resumption_psks(self, number_of_resumption_psks: usize) -> Self

Sets the number_of_resumption_psks property of the MlsGroupCreateConfig.

pub fn use_ratchet_tree_extension( self, use_ratchet_tree_extension: bool ) -> Self

Sets the use_ratchet_tree_extension property of the MlsGroupCreateConfig.

pub fn required_capabilities( self, required_capabilities: RequiredCapabilitiesExtension ) -> Self

Sets the required_capabilities property of the MlsGroupCreateConfig.

pub fn sender_ratchet_configuration( self, sender_ratchet_configuration: SenderRatchetConfiguration ) -> Self

Sets the sender_ratchet_configuration property of the MlsGroupCreateConfig. See SenderRatchetConfiguration for more information.

pub fn lifetime(self, lifetime: Lifetime) -> Self

Sets the lifetime property of the MlsGroupCreateConfig.

pub fn crypto_config(self, config: CryptoConfig) -> Self

Sets the crypto_config property of the MlsGroupCreateConfig.

pub fn external_senders( self, external_senders: ExternalSendersExtension ) -> Self

Sets the external_senders property of the MlsGroupCreateConfig.

pub fn build(self) -> MlsGroupCreateConfig

Finalizes the builder and retursn an [MlsGroupCreateConfig].