openmls::group

Struct StagedCommit

pub struct StagedCommit {
    pub staged_proposal_queue: ProposalQueue,
    /* private fields */
}

Available on 32-bit or 64-bit only.

Contains the changes from a commit to the group state.

Fields
staged_proposal_queue: ProposalQueue

A queue containing the proposals associated with the commit.

Implementations

impl StagedCommit

pub fn add_proposals(&self) -> impl Iterator<Item = QueuedAddProposal<'_>>

Returns the Add proposals that are covered by the Commit message as in iterator over QueuedAddProposal.

pub fn remove_proposals(&self) -> impl Iterator<Item = QueuedRemoveProposal<'_>>

Returns the Remove proposals that are covered by the Commit message as in iterator over QueuedRemoveProposal.

pub fn update_proposals(&self) -> impl Iterator<Item = QueuedUpdateProposal<'_>>

Returns the Update proposals that are covered by the Commit message as in iterator over QueuedUpdateProposal.

pub fn psk_proposals(&self) -> impl Iterator<Item = QueuedPskProposal<'_>>

Returns the PresharedKey proposals that are covered by the Commit message as in iterator over QueuedPskProposal.

pub fn queued_app_ephemeral_proposals( &self, ) -> impl Iterator<Item = QueuedAppEphemeralProposal<'_>>

Available on crate feature extensions-draft-08 only.

Returns the AppEphemeral proposals that are covered by the Commit message as an iterator over QueuedAppEphemeralProposal.

pub fn app_data_update_proposals( &self, ) -> impl Iterator<Item = QueuedAppDataUpdateProposal<'_>>

Available on crate feature extensions-draft-08 only.

Returns the AppDataUpdate proposals that are covered by the Commit message as an iterator over QueuedAppDataUpdateProposal.

pub fn queued_proposals(&self) -> impl Iterator<Item = &QueuedProposal>

Returns an iterator over all QueuedProposals.

pub fn update_path_leaf_node(&self) -> Option<&LeafNode>

Returns the leaf node of the (optional) update path.

pub fn credentials_to_verify(&self) -> impl Iterator<Item = &Credential>

Returns the credentials that the caller needs to verify are valid.

pub fn self_removed(&self) -> bool

Returns true if the member was removed through a proposal covered by this Commit message and false otherwise.

pub fn group_context(&self) -> &GroupContext

Returns the GroupContext of the staged commit state.

pub fn epoch_authenticator(&self) -> Option<&EpochAuthenticator>

Returns the EpochAuthenticator of the staged commit state if the owner of the originating group state is a member of the group. Returns None otherwise.