Struct openmls::key_packages::KeyPackage

pub struct KeyPackage { /* fields omitted */ }

The key package struct.

Implementations

impl KeyPackage

pub fn verify(
&self,
backend: &impl OpenMlsCryptoProvider
) -> Result<(), KeyPackageVerifyError>

Verify that this key package is valid:

verify that the signature on this key package is valid

verify that all mandatory extensions are present

make sure that the lifetime is valid Returns Ok(()) if all checks succeed and KeyPackageError otherwise

pub fn external_key_id(&self) -> Result<&[u8], ExtensionError>

Get the external ID of this key package as byte slice. See ExternalKeyIdExtension for more details on the external key ID extension.

Returns a ExtensionError if no external key ID extension is present.

pub fn extensions(&self) -> &[Extension]

Get a reference to the extensions of this key package.

pub fn check_extension_support(
&self,
required_extensions: &[ExtensionType]
) -> Result<(), KeyPackageExtensionSupportError>

Check whether the this key package supports all the required extensions in the provided list.

pub fn credential(&self) -> &Credential

Get a reference to the Credential.

pub fn hash_ref(
&self,
backend: &impl OpenMlsCrypto
) -> Result<KeyPackageRef, LibraryError>

Compute the KeyPackageRef of this KeyPackage. The KeyPackageRef is used to identify a member in a group (leaf in the tree) within MLS.

pub fn ciphersuite(&self) -> Ciphersuite

Get the Ciphersuite.