Struct openmls::key_packages::KeyPackageBundle

pub struct KeyPackageBundle { /* private fields */ }

A KeyPackageBundle contains a KeyPackage, the corresponding private key, and a leaf secret.

Implementations

impl KeyPackageBundle

pub fn new(
ciphersuites: &[Ciphersuite],
credential_bundle: &CredentialBundle,
backend: &impl OpenMlsCryptoProvider,
extensions: Vec<Extension>
) -> Result<Self, KeyPackageBundleNewError>

Create a new KeyPackageBundle with a fresh key pair. This key package will have the default MLS version. Use KeyPackageBundle::new_with_version to get a key package bundle for a specific MLS version.

Note that the capabilities extension gets added automatically, based on the configuration.

Returns a new KeyPackageBundle or a KeyPackageBundleNewError.

pub fn new_with_version(
version: ProtocolVersion,
ciphersuites: &[Ciphersuite],
backend: &impl OpenMlsCryptoProvider,
credential_bundle: &CredentialBundle,
extensions: Vec<Extension>
) -> Result<Self, KeyPackageBundleNewError>

Create a new KeyPackageBundle with

a fresh key pair

the provided MLS version

the first ciphersuite in the ciphersuites slice

the provided extensions

Note that the capabilities extension gets added automatically, based on the configuration.

Returns a new KeyPackageBundle or a KeyPackageBundleNewError.

pub fn key_package(&self) -> &KeyPackage

Get a reference to the public part of this bundle, i.e. the KeyPackage.

pub fn key_pair(&self) -> (&VLBytes, &HpkePublicKey)

Get a reference to the HPKE key pair.

pub fn hpke_key_pair(&self) -> HpkeKeyPair

Get a reference to the HPKE key pair.

pub fn into_parts(self) -> (KeyPackage, Vec<u8>)

Separates the bundle into the KeyPackage and the HPKE private key as raw byte vectors.