Module openmls::key_store

A storage solution for cryptographic key material used in OpenMLS groups.

This module provides access to the KeyStore struct, which manages the storage of CredentialBundle and KeyPackageBundle instances for use in one or more ManagedGroup instances. The development of this module is tracked in #337, which also includes a roadmap.

Key Store API

All functions accessible via the KeyStore API are thread safe, allowing multiple concurrent read locks, on any category of stored key material (e.g. CredentialBundle instances, “init” KeyPackageBundle instance).

CredentialBundle Instances

The API of the KeyStore allows the generation of Credential instances via generate_credential_bundle, such that it stores the corresponding CredentialBundle. After storing them, references to CredentialBundle instances can be retrieved via get_credential_bundle using the SignaturePublicKey of the corresponding Credential as index.

Init KeyPackageBundle Instances

Similarly, the KeyStore can generate “init” KeyPackage instances via generate_key_package_bundle and store the corresponding KeyPackageBundle. Init KeyPackage instances are meant to be published so other parties can use them to add the publishing party to groups.

KeyPackageBundle vs. CredentialBundle Ownership

In contrast to the functions providing access to CredentialBundle instances, the function to retrieve KeyPackageBundle instances deletes them from the KeyStore. This is because each ManagedGroup currently owns the KeyPackageBundle in its leaf, so upon creation of the group, it needs to consume a KeyPackageBundle instance. Note, that in contrast to CredentialBundle instances, KeyPackageBundle instances should not be used across groups.

This design is temporary and only until the ManagedGroup is refactored to access its KeyPackageBundle via the KeyStore. Once this is the case, the take_key_package_bundle will be deprecated in favor of a get_key_package_bundle, which only returns a reference to the KeyPackageBundle. The issue is tracked as part of the KeyStore roadmap in issue #337.

Example

A simple example for the generation and the retrieval of a CredentialBundle and a KeyPackageBundle.

use openmls::prelude::*;

let key_store = KeyStore::default();

let ciphersuite_name = CiphersuiteName::MLS10_128_DHKEMX25519_AES128GCM_SHA256_Ed25519;

// Generate a credential bundle with the matching signature scheme.
let alice_credential = key_store
    .generate_credential_bundle(
        "Alice".into(),
        CredentialType::Basic,
        SignatureScheme::from(ciphersuite_name),
    )
    .unwrap();

// Generate a key package bundle.
let alice_key_package = key_store
    .generate_key_package_bundle(&[ciphersuite_name], &alice_credential, vec![])
    .unwrap();

// Create a group with the previously generated credential and key package.
let managed_group_config = ManagedGroupConfig::new(
    HandshakeMessageFormat::Plaintext,
    UpdatePolicy::default(),
    0,
    0,
    false, // use_ratchet_tree_extension
    ManagedGroupCallbacks::default(),
);

let alice_group = ManagedGroup::new(
    &key_store,
    &managed_group_config,
    GroupId::from_slice(b"Test Group"),
    &alice_key_package.hash(),
)
.unwrap();

Future Work

A more detailed roadmap can be found in issue #337, but generally the plan is to

move ownership of all KeyPackageBundle instances into the KeyStore and change the API from take_ to get_

keep track of which key material is (still) in use and where, as well as if it expires and when

extend the API to allow for deletion of unused key material, expired or otherwise

add a persistence layer

Re-exports

pub use errors::KeyStoreError;

Modules

errors

Structs

CbGuard
This guard struct for a CredentialBundle implements Deref, such that the underlying RwLock<CredentialBundle> can be obtained for read or write access to the credential.

KeyStore
The KeyStore contains CredentialBundles and KeyPackageBundles and makes them available via information in the corresponding Credential and KeyPackage instances.