Struct openmls::key_store::KeyStore

pub struct KeyStore {
    credential_bundles: RwLock<HashMap<SignaturePublicKey, CredentialBundle>>,
    init_key_package_bundles: RwLock<HashMap<Vec<u8>, KeyPackageBundle>>,
}

The KeyStore contains CredentialBundles and KeyPackageBundles and makes them available via information in the corresponding Credential and KeyPackage instances.

Fields
credential_bundles: RwLock<HashMap<SignaturePublicKey, CredentialBundle>>init_key_package_bundles: RwLock<HashMap<Vec<u8>, KeyPackageBundle>>
Implementations

impl KeyStore

pub fn take_key_package_bundle(
&self,
kp_hash: &[u8]
) -> Option<KeyPackageBundle>

Retrieve a KeyPackageBundle from the key store given the hash of the corresponding KeyPackage. This removes the KeyPackageBundle from the store. Returns an error if no KeyPackageBundle can be found corresponding to the given KeyPackage hash.

pub(crate) fn get_credential_bundle<'key_store>(
&'key_store self,
signature_public_key: &'key_store SignaturePublicKey
) -> Option<CbGuard<'_>>

Retrieve a CbGuard from the key store given the SignaturePublicKey of the corresponding Credential. The CbGuard can be dereferenced to obtain an RwLock on the desired CredentialBundle. Returns an error if no CredentialBundle can be found corresponding to the given SignaturePublicKey.

pub fn generate_key_package_bundle(
&self,
ciphersuites: &[CiphersuiteName],
credential: &Credential,
extensions: Vec<Box<dyn Extension>>
) -> Result<KeyPackage, KeyStoreError>

Generate a fresh KeyPackageBundle with the given parameters, store it in the KeyStore and return the corresponding KeyPackage. Throws an error if no CredentialBundle can be found in the KeyStore corresponding to the given Credential or if an error occurred during the creation of the KeyPackageBundle.

pub fn generate_credential_bundle(
&self,
identity: Vec<u8>,
credential_type: CredentialType,
signature_scheme: SignatureScheme
) -> Result<Credential, KeyStoreError>

Generate a fresh CredentialBundle with the given parameters and store it in the KeyStore. Returns the corresponding Credential or an error if the creation of the CredentialBundle fails.

pub fn get_leaf_secret(&self, kp_hash: &[u8]) -> Secret

This function allows us to get hold of the next leaf_secret in the TreeKEM test vector.