openmls

Module messages

Available on 32-bit or 64-bit only.

Messages

This module defines types and logic for Commit and Welcome messages, as well as Proposals and group info used in External Commits.

Modules

external_proposals

External Proposals

group_info

This module contains all types related to group info handling.

proposals

Proposals

proposals_in

Proposals

Structs

ConfirmationTag

Confirmation tag field of PublicMessage. For type safety this is a wrapper around a Mac.

EncryptedGroupSecrets

EncryptedGroupSecrets

Welcome

Welcome message

Enums

GroupSecretsError

Error related to group secrets.