Module openmls::messages::group_info

This module contains all types related to group info handling.

Structs

GroupInfo

GroupInfo

VerifiableGroupInfo

A type that represents a group info of which the signature has not been verified. It implements the Verifiable trait and can be turned into a group info by calling verify(...) with the Credential corresponding to the CredentialBundle of the signer. When receiving a serialized group info, it can only be deserialized into a VerifiableGroupInfo, which can then be turned into a group info as described above.