Struct openmls::messages::group_info::VerifiableGroupInfo

pub struct VerifiableGroupInfo { /* private fields */ }

A type that represents a group info of which the signature has not been verified. It implements the Verifiable trait and can be turned into a group info by calling verify(...) with the signature key of the Credential. When receiving a serialized group info, it can only be deserialized into a VerifiableGroupInfo, which can then be turned into a group info as described above.

Implementations

impl VerifiableGroupInfo

pub fn new( group_context: GroupContext, extensions: Extensions, confirmation_tag: ConfirmationTag, signer: LeafNodeIndex, signature: Signature ) -> Self

Create a new VerifiableGroupInfo from its contents.

pub fn ciphersuite(&self) -> Ciphersuite

Get (unverified) ciphersuite of the verifiable group info.

Note: This method should only be used when necessary to verify the group info signature.