openmls::messages

Module proposals_in

Available on 32-bit or 64-bit only.

Proposals

This module defines all the different types of Proposals.

Structs

AddProposalIn

Add Proposal.

GroupContextExtensionProposalIn

GroupContext Extension Proposal.

UpdateProposalIn

Update Proposal.

Enums

ProposalIn

Proposal.

ProposalOrRefIn

Type of Proposal, either by value or by reference.