Module openmls::messages::public_group_state

This module contains the types and implementations for the PublicGroupState and PublicGroupStateTbs structs of the MLS spec. The PublicGroupState implements type-enforced verification using the set of traits defined in the signable module in the same way as the MlsPlaintext.

Structs

PublicGroupState

PublicGroupState as defined in the MLS specification as follows:

VerifiablePublicGroupState

The VerifiablePublicGroupState represents a PublicGroupState of which the signature has not been verified. It implements the Verifiable trait and can thus be turned into a PublicGroupState by calling verify(...) with the Credential corresponding to the CredentialBundle of the signer. When receiving a serialized PublicGroupState, it can thus only be deserialized into a VerifiablePublicGroupState, which can then be turned into a PublicGroupState as described above.