Struct openmls::prelude::KeyPackage

pub struct KeyPackage { /* fields omitted */ }

Implementations

impl KeyPackage

Public KeyPackage functions.

pub fn verify(&self) -> Result<(), KeyPackageError>

Verify that this key package is valid:

verify that the signature on this key package is valid

verify that all mandatory extensions are present

make sure that the lifetime is valid Returns Ok(()) if all checks succeed and KeyPackageError otherwise

pub fn hash(&self) -> Vec<u8>

Compute the hash of the encoding of this key package.

pub fn key_id(&self) -> Result<&[u8], KeyPackageError>

Get the ID of this key package as byte slice. Returns an error if no Key ID extension is present.

pub fn extensions(&self) -> &[Extension]

Get a reference to the extensions of this key package.

pub fn credential(&self) -> &Credential

Get a reference to the credential.

impl KeyPackage

Crate visible KeyPackage functions.

pub fn ciphersuite_name(&self) -> CiphersuiteName

Get the CiphersuiteName.