openmls::prelude

Struct TlsByteVecU8

pub struct TlsByteVecU8 { /* private fields */ }

Available on 32-bit or 64-bit only.

Implementations

impl TlsByteVecU8

pub fn new(vec: Vec<u8>) -> TlsByteVecU8

Create a new TlsVec from a Rust Vec.

pub fn from_slice(slice: &[u8]) -> TlsByteVecU8
where u8: Clone,

Create a new TlsVec from a slice.

pub fn len(&self) -> usize

Get the length of the vector.

pub fn as_slice(&self) -> &[u8]

Get a slice to the raw vector.

pub fn is_empty(&self) -> bool

Check if the vector is empty.

pub fn into_vec(self) -> Vec<u8>

Get the underlying vector and consume this.

pub fn push(&mut self, value: u8)

Add an element to this.

pub fn pop(&mut self) -> Option<u8>

Remove the last element.

pub fn remove(&mut self, index: usize) -> u8

Remove the element at index.

pub fn get(&self, index: usize) -> Option<&u8>

Returns a reference to an element or subslice depending on the type of index. XXX: implement SliceIndex instead

pub fn iter(&self) -> Iter<'_, u8>

Returns an iterator over the slice.

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&u8) -> bool,

Retains only the elements specified by the predicate.

pub fn len_len() -> usize

Get the number of bytes used for the length encoding.