Struct openmls::prelude::TlsVecU8

pub struct TlsVecU8<T> { /* private fields */ }

Implementations

impl<T> TlsVecU8<T>

pub fn new(vec: Vec<T>) -> TlsVecU8<T>

Create a new TlsVec from a Rust Vec.

pub fn from_slice(slice: &[T]) -> TlsVecU8<T>
where T: Clone,

Create a new TlsVec from a slice.

pub fn len(&self) -> usize

Get the length of the vector.

pub fn as_slice(&self) -> &[T]

Get a slice to the raw vector.

pub fn is_empty(&self) -> bool

Check if the vector is empty.

pub fn into_vec(self) -> Vec<T>

Get the underlying vector and consume this.

pub fn push(&mut self, value: T)

Add an element to this.

pub fn pop(&mut self) -> Option<T>

Remove the last element.

pub fn remove(&mut self, index: usize) -> T

Remove the element at index.

pub fn get(&self, index: usize) -> Option<&T>

Returns a reference to an element or subslice depending on the type of index. XXX: implement SliceIndex instead

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over the slice.

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

pub fn len_len() -> usize

Get the number of bytes used for the length encoding.