Trait openmls::prelude::OpenMlsKeyStore

pub trait OpenMlsKeyStore: Send + Sync {
    type Error: Error + Debug + PartialEq<Self::Error>;

    fn read_epoch_keys<V>(
        &self,
        identity: &[u8],
        group_id: &[u8],
        epoch: u64
    ) -> Vec<V, Global>
    where
        V: FromKeyStoreValue;
    fn store_epoch_keys<V>(
        &self,
        identity: &[u8],
        group_id: &[u8],
        epoch: u64,
        encryption_keys: &[V]
    ) -> Result<(), Self::Error>
    where
        V: ToKeyStoreValue;
    fn delete_epoch_keys(
        &self,
        identity: &[u8],
        group_id: &[u8],
        epoch: u64
    ) -> Result<(), Self::Error>;
    fn store<V>(&self, k: &[u8], v: &V) -> Result<(), Self::Error>
    where
        V: ToKeyStoreValue,
        Self: Sized;
    fn read<V>(&self, k: &[u8]) -> Option<V>
    where
        V: FromKeyStoreValue,
        Self: Sized;
    fn delete(&self, k: &[u8]) -> Result<(), Self::Error>;
}

The Key Store trait

Required Associated Types

type Error: Error + Debug + PartialEq<Self::Error>

The error type returned by the OpenMlsKeyStore.

Required Methods

fn read_epoch_keys<V>(
&self,
identity: &[u8],
group_id: &[u8],
epoch: u64
) -> Vec<V, Global>where
V: FromKeyStoreValue,

Load all encryption keys associated with the given client identity (i.e. a byte vector uniquely identifying the client), epoch and group ID from the key store.

fn store_epoch_keys<V>(
&self,
identity: &[u8],
group_id: &[u8],
epoch: u64,
encryption_keys: &[V]
) -> Result<(), Self::Error>where
V: ToKeyStoreValue,

Store all encryption keys associated with the given client identity (i.e. a byte vector uniquely identifying the client), epoch and group ID in the key store.

fn delete_epoch_keys(
&self,
identity: &[u8],
group_id: &[u8],
epoch: u64
) -> Result<(), Self::Error>

Delete all encryption keys associated with the given client identity (i.e. a byte vector uniquely identifying the client), epoch and group ID from the key store.

fn store<V>(&self, k: &[u8], v: &V) -> Result<(), Self::Error>where
V: ToKeyStoreValue,
Self: Sized,

Store a value v that implements the [ToKeyStoreValue] trait for serialization for ID k.

Returns an error if storing fails.

fn read<V>(&self, k: &[u8]) -> Option<V>where
V: FromKeyStoreValue,
Self: Sized,

Read and return a value stored for ID k that implements the [FromKeyStoreValue] trait for deserialization.

Returns None if no value is stored for k or reading fails.

fn delete(&self, k: &[u8]) -> Result<(), Self::Error>

Delete a value stored for ID k.

Returns an error if storing fails.