Trait openmls::prelude::OpenMlsProvider

pub trait OpenMlsProvider {
    type CryptoProvider: OpenMlsCrypto;
    type RandProvider: OpenMlsRand;
    type KeyStoreProvider: OpenMlsKeyStore;

    // Required methods
    fn crypto(&self) -> &Self::CryptoProvider;
    fn rand(&self) -> &Self::RandProvider;
    fn key_store(&self) -> &Self::KeyStoreProvider;
}

The OpenMLS Crypto Provider Trait

An implementation of this trait must be passed in to the public OpenMLS API to perform randomness generation, cryptographic operations, and key storage.

Required Associated Types

type CryptoProvider: OpenMlsCrypto

type RandProvider: OpenMlsRand

type KeyStoreProvider: OpenMlsKeyStore

Required Methods

fn crypto(&self) -> &Self::CryptoProvider

Get the crypto provider.

fn rand(&self) -> &Self::RandProvider

Get the randomness provider.

fn key_store(&self) -> &Self::KeyStoreProvider

Get the key store provider.