openmls

Module schedule

Available on 32-bit or 64-bit only.

Key Schedule

This module defines types and implementations for key schedule operations. It provides the EpochAuthenticator and ResumptionPskSecret types.

Internal Documentation

The key schedule is described in Section 8 of the MLS specification. It evolves in epochs, with new key material injected in each epoch.

The key schedule flow (from Section 8 of the MLS specification) is as follows:

Re-exports

pub use psk::ExternalPsk;

pub use psk::PreSharedKeyId;

pub use psk::Psk;

Modules

errors

Key schedule errors.

psk

Preshared keys.

tests_and_katstest-utils

Structs

ApplicationExportSecretextensions-draft-08

A secret that we can derive secrets from, that are used outside of OpenMLS. In contrast to [ExporterSecret], the [ApplicationExportSecret] is not persisted. It can be deleted after use to achieve forward secrecy.

EpochAuthenticator

A secret that can be used among members to make sure everyone has the same group state.

ResumptionPskSecret

A group secret that can be used among members to prove that a member was part of a group in a given epoch.

Enums

PprfErrorextensions-draft-08

Error evaluating the PPRF at the given input.