Module openmls::schedule::psk

Pre shared keys.

enum {
    reserved(0),
    external(1),
    reinit(2),
    branch(3),
    (255)
  } PSKType;

  struct {
    PSKType psktype;
    select (PreSharedKeyID.psktype) {
      case external:
        opaque psk_id<0..255>;

      case reinit:
        opaque psk_group_id<0..255>;
        uint64 psk_epoch;

      case branch:
        opaque psk_group_id<0..255>;
        uint64 psk_epoch;
    }
    opaque psk_nonce<0..255>;
  } PreSharedKeyID;

  struct {
      PreSharedKeyID psks<0..2^16-1>;
  } PreSharedKeys;

Structs

BranchPsk
Branch PSK

ExternalPsk
External PSK.

ExternalPskBundle
External PSK Bundle. This contains the secret part of the PSK as well as the public part that is used as a marker for injection into the key schedule.

PreSharedKeyId
A PreSharedKeyID is used to uniquely identify the PSKs that get injected in the key schedule.

PreSharedKeys
PreSharedKeys is a vector of PreSharedKeyIDs. struct { PreSharedKeyID psks<0..2^16-1>; } PreSharedKeys;

PskLabel
PskLabel is used in the final concatentation of PSKs before they are injected in the key schedule. struct { PreSharedKeyID id; uint16 index; uint16 count; } PSKLabel;

PskSecret
This contains the psk-secret calculated from the PSKs contained in a Commit or a PreSharedKey proposal.

ReinitPsk
ReInit PSK.

Enums

Psk
PSK enum that can contain the different PSK types

PskType
Type of PSK.