openmls::storage
Trait OpenMlsProvider

pub trait OpenMlsProvider: OpenMlsProvider<StorageProvider = Self::Storage> {
    type Storage: StorageProvider<Error = Self::StorageError>;
    type StorageError: Error;
}

A convenience trait for the OpenMLS provider that defines the storage provider for the current version of storage. Throughout the code, this one should be used instead of openmls_traits::OpenMlsProvider.

Required Associated Types

type Storage: StorageProvider<Error = Self::StorageError>

The storage to use

type StorageError: Error

The storage error type