openmls

Module test_utils

Available on (32-bit or 64-bit) and (crate features test-utils) only.

Test utilities

Modules

frankenstein

This module contains the Frankenstein test utilities.

single_group_test_framework

storage_state

test_framework

This crate provides a framework to set up clients and groups using the OpenMLS mls_group API. To use the framework, start by creating a new TestSetup with a number of clients. After that, create_clients has to be called before the the TestSetup can be used.

vector_converter

Helper mod that converts a objects that implement FromIterator<,> (like a HashMap or a BTreeMap) into a vector of tuples and vice versa.

Structs

HpkeKeyPair

Helper holding a (private, public) key pair as byte vectors.

Enums

Ciphersuite

MLS ciphersuites.

Traits

OpenMlsProvider

The OpenMLS Crypto Provider Trait

StorageProviderTrait

StorageProvider describes the storage backing OpenMLS and persists the state of OpenMLS groups.

Functions

bytes_to_hex

Convert bytes to a hex string.

hex_to_bytes

Convert a hex string to a byte vector.

hex_to_bytes_option

Convert a hex string to a byte vector. If the input is None, this returns an empty vector.

random_u8

random_u32

random_u64

Type Aliases

OpenMlsLibcruxlibcrux-provider and neither WebAssembly, nor x86 and Windows

OpenMlsRustCrypto