openmls::test_utils

Module frankenstein

Available on (32-bit or 64-bit) and (crate features test-utils) only.

This module contains the Frankenstein test utilities.

The Frankenstein test utilities are used to build and manipulate test structures in a way that is not possible with the public API. This is useful for testing and fuzzing.

Structs

FrankenAddProposal

FrankenAppEphemeralProposalextensions-draft-08

FrankenApplicationIdExtension

FrankenAuthenticatedContentTbm

FrankenCapabilities

FrankenCommit

FrankenCredential

FrankenCustomProposal

FrankenEncryptedGroupSecrets

FrankenExternalInitProposal

FrankenExternalPsk

FrankenExternalPubExtension

FrankenExternalSender

FrankenExternalSendersExtension

FrankenFramedContent

FrankenFramedContentAuthData

FrankenFramedContentTbs

FrankenHpkeCiphertext

FrankenKeyPackage

FrankenKeyPackageTbs

FrankenLeafNode

FrankenLeafNodePayload

FrankenLeafNodeTbs

FrankenLifetime

FrankenMessageRange

FrankenMlsMessage

FrankenParentNode

FrankenPreSharedKeyId

FrankenPreSharedKeyProposal

FrankenPrivateMessage

FrankenPublicMessage

FrankenRatchetTreeExtension

FrankenReInitProposal

FrankenRemoveProposal

FrankenRequiredCapabilitiesExtension

FrankenResumptionPsk

FrankenSignContent

FrankenTreePosition

FrankenUpdatePathIn

FrankenUpdatePathNode

FrankenUpdateProposal

FrankenWelcome

Enums

FrankenContentType

FrankenExtension

FrankenExtensionType

FrankenFramedContentBody

FrankenLeafNodeSource

FrankenMlsMessageBody

FrankenNode

FrankenProposal

FrankenProposalOrRef

FrankenProposalType

FrankenPsk

FrankenResumptionPskUsage

FrankenSender

Functions

compute_membership_tag

Computes a valid membership tag for the provided content.

sign_with_label

Implements the “sign with label” function of the spec.