openmls::test_utils::restricted_provider

Struct RestrictedCrypto

pub struct RestrictedCrypto { /* private fields */ }

Available on (32-bit or 64-bit) and (crate features test-utils) only.

A crypto provider that performs all operations via [RustCrypto] but only claims support for an explicit allowlist of ciphersuites.