openmls::test_utils

Module single_group_test_framework

Available on (32-bit or 64-bit) and (crate features test-utils) only.

Modules

vector_converter

Helper mod that converts a objects that implement FromIterator<,> (like a HashMap or a BTreeMap) into a vector of tuples and vice versa.

Structs

AddMemberConfig

CorePartyState

Struct representing a party’s global state

GroupState

All of the state for a group and its members

HpkeKeyPair

Helper holding a (private, public) key pair as byte vectors.

MemberState

Represents a group member’s MlsGroup instance and pre-group state

PreGroupPartyState

Struct representing a party’s state before joining a group

PreGroupPartyStateBuilder

Enums

Ciphersuite

MLS ciphersuites.

Traits

OpenMlsProvider

The OpenMLS Crypto Provider Trait

StorageProviderTrait

StorageProvider describes the storage backing OpenMLS and persists the state of OpenMLS groups.

Functions

generate_credential

Helper function for generating a credential.

random_u8

random_u32

random_u64

Type Aliases

GroupError