Module openmls::tree

Modules

codec

hash_input
7.5. Tree Hashes

index

node

path_keys
A data structure holding HPKE key pairs for a path in the tree.

private_tree
A data structure holding information about the leaf node in the tree that belongs to the current client.

secret_tree

sender_ratchet

treemath

Structs

RatchetTree
The ratchet tree.

UpdatePath
7.7. Update Paths

UpdatePathNode
7.7. Update Paths

Enums

TreeError
These are errors the RatchetTree can return.

Type Definitions

InvitationList