Struct openmls::tree::RatchetTree

pub struct RatchetTree {
    ciphersuite: &'static Ciphersuite,
    pub nodes: Vec<Node>,
    private_tree: PrivateTree,
}

The ratchet tree.

Fields
ciphersuite: &'static Ciphersuite

The ciphersuite used in this tree.

nodes: Vec<Node>

All nodes in the tree. Note that these only hold public values. Private values are stored in the private_tree.

private_tree: PrivateTree

This holds all private values in the tree. See PrivateTree for details.

Implementations

impl RatchetTree

pub(crate) fn new(
ciphersuite: &'static Ciphersuite,
kpb: KeyPackageBundle
) -> RatchetTree

Create a new empty RatchetTree.

pub(crate) fn new_from_public_tree(ratchet_tree: &RatchetTree) -> Self

Create a new RatchetTree by cloning the public tree nodes from another tree and an empty PrivateTree

pub(crate) fn new_from_nodes(
ciphersuite: &'static Ciphersuite,
kpb: KeyPackageBundle,
node_options: &[Option<Node>]
) -> Result<RatchetTree, TreeError>

Generate a new RatchetTree from Nodes with the client's key package bundle kpb.

pub(crate) fn private_tree_mut(&mut self) -> &mut PrivateTree

Return a mutable reference to the PrivateTree.

fn tree_size(&self) -> NodeIndex

pub fn public_key_tree(&self) -> Vec<Option<&Node>>

Get a vector with all nodes in the tree, containing None for blank nodes.

pub fn public_key_tree_copy(&self) -> Vec<Option<Node>>

Get a vector with a copy of all nodes in the tree, containing None for blank nodes.

pub(crate) fn leaf_count(&self) -> LeafIndex

fn resolve(&self, index: NodeIndex) -> Vec<NodeIndex>

pub(crate) fn get_own_node_index(&self) -> NodeIndex

Get the index of the own node.

fn get_own_key_package_ref(&self) -> &KeyPackage

Get a reference to the own key package.

fn get_own_key_package_ref_mut(&mut self) -> &mut KeyPackage

Get a mutable reference to the own key package.

fn blank_member(&mut self, index: NodeIndex)

fn free_leaves(&self) -> Vec<NodeIndex>

pub(crate) fn update_path(
&mut self,
sender: LeafIndex,
update_path: &UpdatePath,
group_context: &[u8]
) -> Result<Secret, TreeError>

7.7. Update Paths

Update the path for incoming commits.

The path contains a public key and encrypted secret value for all intermediate nodes in the path above the leaf. The path is ordered from the closest node to the leaf to the root; each node MUST be the parent of its predecessor.

pub(crate) fn replace_private_tree(
&mut self,
ciphersuite: &Ciphersuite,
key_package_bundle: &KeyPackageBundle,
group_context: &[u8]
) -> Result<Secret, TreeError>

Update the private tree with the new KeyPackageBundle.

pub(crate) fn refresh_private_tree(
&mut self,
ciphersuite: &Ciphersuite,
credential_bundle: &CredentialBundle,
group_context: &[u8]
) -> (Secret, UpdatePath, Vec<Secret>, KeyPackageBundle)

Update the private tree.

fn replace_private_tree_(
&mut self,
ciphersuite: &Ciphersuite,
key_package_bundle: &KeyPackageBundle,
group_context: &[u8],
with_update_path: bool
) -> Option<UpdatePath>

Replace the private tree with a new one based on the key_package_bundle.

fn encrypt_to_copath(
&self,
public_keys: Vec<HPKEPublicKey>,
group_context: &[u8]
) -> Result<Vec<UpdatePathNode>, TreeError>

Encrypt the path secrets to the co path and return the update path.

fn merge_direct_path_keys(
&mut self,
direct_path: &UpdatePath,
path: Vec<NodeIndex>
) -> Result<(), TreeError>

Merge public keys from a direct path to this tree along the given path.

pub(crate) fn validate_public_keys(
&self,
public_keys: &[HPKEPublicKey],
path: &[NodeIndex]
) -> Result<(), TreeError>

Validates that the public_keys matches the public keys in the tree along path

pub(crate) fn merge_public_keys(
&mut self,
public_keys: &[HPKEPublicKey],
path: &[NodeIndex]
) -> Result<(), TreeError>

Merges public_keys into the tree along the path

pub(crate) fn add_nodes(
&mut self,
new_kps: &[&KeyPackage]
) -> Vec<(NodeIndex, Credential)>

Add nodes for the provided key packages.

pub fn apply_proposals(
&mut self,
proposal_id_list: &[ProposalID],
proposal_queue: ProposalQueue,
updates_key_package_bundles: &[KeyPackageBundle]
) -> Result<(bool, bool, Vec<(NodeIndex, AddProposal)>), TreeError>

Applies a list of proposals from a Commit to the tree. proposal_id_list corresponds to the proposals field of a Commit proposal_queue is the queue of proposals received or sent in the current epoch updates_key_package_bundles is the list of own KeyPackageBundles corresponding to updates or commits sent in the current epoch

fn trim_tree(&mut self)

Trims the tree from the right when there are empty leaf nodes

pub fn compute_tree_hash(&self) -> Vec<u8>

Computes the tree hash

pub fn compute_parent_hash(&mut self, index: NodeIndex) -> Vec<u8>

Computes the parent hash

pub fn verify_integrity(
ciphersuite: &Ciphersuite,
nodes: &[Option<Node>]
) -> bool

Verifies the integrity of a public tree