Module openmls::tree::tests_and_kats::kats::kat_encryption

Known Answer Tests for encrypting to tree nodes

See https://github.com/mlswg/mls-implementations/blob/master/test-vectors.md for more description on the test vectors.

Parameters:

Ciphersuite

Number of leaves

Number of generations

Format:

{
  "cipher_suite": /* uint16 */,
  "n_leaves": /* uint32 */,
  "encryption_secret": /* hex-encoded binary data */,
  "sender_data_secret": /* hex-encoded binary data */,
  "sender_data_info": {
    "ciphertext": /* hex-encoded binary data */,
    "key": /* hex-encoded binary data */,
    "nonce": /* hex-encoded binary data */,
  },
  "leaves": [
    {
      "generations": /* uint32 */,
      "handshake": [ /* array with `generations` handshake keys and nonces */
        {
          "key": /* hex-encoded binary data */,
          "nonce": /* hex-encoded binary data */,
          "plaintext": /* hex-encoded binary data */
          "ciphertext": /* hex-encoded binary data */
        },
        ...
      ],
      "application": [ /* array with `generations` application keys and nonces */
        {
          "key": /* hex-encoded binary data */,
          "nonce": /* hex-encoded binary data */,
          "plaintext": /* hex-encoded binary data */
          "ciphertext": /* hex-encoded binary data */
        },
        ...
      ]
    }
  ]
}

Verification:

For all N entries in the leaves and all generations j

leaves[N].handshake[j].key = handshake_ratchet_key_[2*N]_[j]

leaves[N].handshake[j].nonce = handshake_ratchet_nonce_[2*N]_[j]

leaves[N].handshake[j].plaintext represents an MlsPlaintext containing a handshake message (Proposal or Commit) from leaf N

leaves[N].handshake[j].ciphertext represents an MlsCiphertext object that successfully decrypts to an MlsPlaintext equivalent to leaves[N].handshake[j].plaintext using the keys for leaf N and generation j.

leaves[N].application[j].key = application_ratchet_key_[2*N]_[j]

leaves[N].application[j].nonce = application_ratchet_nonce_[2*N]_[j]

leaves[N].application[j].plaintext represents an MlsPlaintext containing application data from leaf N

leaves[N].application[j].ciphertext represents an MlsCiphertext object that successfully decrypts to an MlsPlaintext equivalent to leaves[N].handshake[j].plaintext using the keys for leaf N and generation j.

sender_data_info.secret.key = sender_data_key(sender_data_secret, sender_data_info.ciphertext)

sender_data_info.secret.nonce = sender_data_nonce(sender_data_secret, sender_data_info.ciphertext)

The extra factor of 2 in 2*N ensures that only chains rooted at leaf nodes are tested. The definitions of ratchet_key and ratchet_nonce are in the Encryption Keys section of the specification.

Structs

EncryptionTestVector

Enums

EncTestVectorError

Functions

generate_test_vector

run_test_vector