Module openmls::tree::treemath

Enums

TreeMathError

Functions

common_ancestor_index

copath
Copath of a leaf. Ordered from leaf to root.

leaf_count
Returns the number of leaves in a tree

leaf_direct_path
Direct path from a leaf node to the root. Does not include the leaf node but includes the root.

leaf_in_tree

left

level

log2

node_in_tree

node_width

parent

parent_direct_path
Direct path from a parent node to the root. Includes the parent node and the root. Returns an error if the index is not a parent node.

parent_step

right

root

sibling

unsafe_parent