Crate openmls

Available on 32-bit or 64-bit only.

OpenMLS

OpenMLS is a Rust implementation of the Messaging Layer Security (MLS) protocol, as specified in RFC 9420.

It is a software library that can serve as a building block in applications that require end-to-end encryption of messages. It has a safe and easy-to-use interface that hides the complexity of the underlying cryptographic operations.

Supported ciphersuites

MLS_128_DHKEMX25519_AES128GCM_SHA256_Ed25519 (MTI)

MLS_128_DHKEMP256_AES128GCM_SHA256_P256

MLS_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519

Supported platforms

OpenMLS is built and tested on the Github CI for the following rust targets.

x86_64-unknown-linux-gnu

i686-unknown-linux-gnu

x86_64-pc-windows-msvc

i686-pc-windows-msvc

x86_64-apple-darwin

Unsupported, but built on CI

The Github CI also builds (but doesn’t test) the following rust targets.

aarch64-apple-darwin

aarch64-unknown-linux-gnu

aarch64-linux-android

aarch64-apple-ios

aarch64-apple-ios-sim

wasm32-unknown-unknown

armv7-linux-androideabi

x86_64-linux-android

i686-linux-android

OpenMLS supports 32 bit platforms and above.

Cryptography Dependencies

OpenMLS does not implement its own cryptographic primitives. Instead, it relies on existing implementations of the cryptographic primitives used by MLS. There are two different cryptography providers implemented right now. But consumers can bring their own implementation. See traits for more details.

Features

OpenMLS provides the following features

extensions-draft-08: enable features defined in MLS extensions draft-08

fork-resolution: helper functionality for resolving forks.

js: enable compilation to wasm

Developer features

libcrux-provider: enable the libcrux crypto provider dependency

openmls_rust_crypto: enable the rust crypto provider

sqlite-provider: enable the sqlite provider

backtrace: enable backtraces

content-debug: allow printing sensitive content of messages for debugging

crypto-debug: allow printing cryptographic key material for debugging

test-util: test utilities

Working on OpenMLS

For more details when working on OpenMLS itself please see the Developer.md.

Maintenance & Support

OpenMLS is maintained and developed by Phoenix R&D and Cryspen.

Acknowledgements

Zulip graciously provides the OpenMLS community with a “Zulip Cloud Standard” tier Zulip instance.

Quick Start

For a quick start to learn how OpenMLS works here’s the basic code to set up to parties and have them create a group.

use openmls::{prelude::{*,  tls_codec::*}};
use openmls_rust_crypto::{OpenMlsRustCrypto};
use openmls_basic_credential::SignatureKeyPair;

// Define ciphersuite ...
let ciphersuite = Ciphersuite::MLS_128_DHKEMX25519_AES128GCM_SHA256_Ed25519;
// ... and providers for both parties to use.
let sasha_provider = &OpenMlsRustCrypto::default();
let maxim_provider = &OpenMlsRustCrypto::default();

// Now let's create two participants.

// A helper to create and store credentials.
fn generate_credential_with_key(
    identity: Vec<u8>,
    credential_type: CredentialType,
    signature_algorithm: SignatureScheme,
    provider: &impl OpenMlsProvider,
) -> (CredentialWithKey, SignatureKeyPair) {
    let credential = BasicCredential::new(identity);
    let signature_keys =
        SignatureKeyPair::new(signature_algorithm)
            .expect("Error generating a signature key pair.");

    // Store the signature key into the key store so OpenMLS has access
    // to it.
    signature_keys
        .store(provider.storage())
        .expect("Error storing signature keys in key store.");

    (
        CredentialWithKey {
            credential: credential.into(),
            signature_key: signature_keys.public().into(),
        },
        signature_keys,
    )
}

// A helper to create key package bundles.
fn generate_key_package(
    ciphersuite: Ciphersuite,
    provider: &impl OpenMlsProvider,
    signer: &SignatureKeyPair,
    credential_with_key: CredentialWithKey,
) -> KeyPackageBundle {
    // Create the key package
    KeyPackage::builder()
        .build(
            ciphersuite,
            provider,
            signer,
            credential_with_key,
        )
        .unwrap()
}

// First they need credentials to identify them
let (sasha_credential_with_key, sasha_signer) = generate_credential_with_key(
    "Sasha".into(),
    CredentialType::Basic,
    ciphersuite.signature_algorithm(),
    sasha_provider,
);

let (maxim_credential_with_key, maxim_signer) = generate_credential_with_key(
    "Maxim".into(),
    CredentialType::Basic,
    ciphersuite.signature_algorithm(),
    maxim_provider,
);

// Then they generate key packages to facilitate the asynchronous handshakes
// in MLS

// Generate KeyPackages
let maxim_key_package = generate_key_package(ciphersuite, maxim_provider, &maxim_signer, maxim_credential_with_key);

// Now Sasha starts a new group ...
let mut sasha_group = MlsGroup::new(
    sasha_provider,
    &sasha_signer,
    &MlsGroupCreateConfig::default(),
    sasha_credential_with_key,
)
.expect("An unexpected error occurred.");

// ... and invites Maxim.
// The key package has to be retrieved from Maxim in some way. Most likely
// via a server storing key packages for users.
let (mls_message_out, welcome_out, group_info) = sasha_group
    .add_members(sasha_provider, &sasha_signer, core::slice::from_ref(maxim_key_package.key_package()))
    .expect("Could not add members.");

// Sasha merges the pending commit that adds Maxim.
sasha_group
   .merge_pending_commit(sasha_provider)
   .expect("error merging pending commit");

// Sasha serializes the [`MlsMessageOut`] containing the [`Welcome`].
let serialized_welcome = welcome_out
   .tls_serialize_detached()
   .expect("Error serializing welcome");

// Maxim can now de-serialize the message as an [`MlsMessageIn`] ...
let mls_message_in = MlsMessageIn::tls_deserialize(&mut serialized_welcome.as_slice())
   .expect("An unexpected error occurred.");

// ... and inspect the message.
let welcome = match mls_message_in.extract() {
   MlsMessageBodyIn::Welcome(welcome) => welcome,
   // We know it's a welcome message, so we ignore all other cases.
   _ => unreachable!("Unexpected message type."),
};

// Now Maxim can build a staged join for the group in order to inspect the welcome
let maxim_staged_join = StagedWelcome::new_from_welcome(
    maxim_provider,
    &MlsGroupJoinConfig::default(),
    welcome,
    // The public tree is needed and transferred out of band.
    // It is also possible to use the [`RatchetTreeExtension`]
    Some(sasha_group.export_ratchet_tree().into()),
)
.expect("Error creating a staged join from Welcome");

// Finally, Maxim can create the group
let mut maxim_group = maxim_staged_join
    .into_group(maxim_provider)
    .expect("Error creating the group from the staged join");

Modules

ciphersuite

Ciphersuites for MLS

componentextensions-draft-08

Components from the extensions draft.

credentials

Credentials

error

OpenMLS Errors

extensions

Extensions

framing

Message framing

grease

GREASE Support for MLS

group

Group API for MLS

key_packages

Key Packages

messages

Messages

prelude

Single place, re-exporting the most used public functions. Prelude for OpenMLS. Include this to get access to all the public functions of OpenMLS.

prelude_testtest-utils

Single place, re-exporting all structs and functions needed for integration tests Testing prelude for OpenMLS. Include this to get access to all necessary pub(crate) functions of OpenMLS testing.

schedule

Key Schedule

storage

OpenMLS Storage

test_utilstest-utils

Test utilities

treesync

This module implements the ratchet tree component of MLS.

versions

MLS versions

wasmtest-utils