Crate openmls

OpenMLS

High-level API

See ManagedGroup

Low-level API

See MlsGroup

Errors

Each module has an errors.rs defining module specific errors that are used within the crate. This exposes some of the module errors that are publicly relevant. All errors implement the Error trait and PartialEq.

The high-level errors API in error are a different error representation as u16 for C FFI APIs.

Re-exports

pub use crate::tree::node;

Modules

ciphersuite
Ciphersuites for MLS

codec

config
This config contains all structs, enums and functions to configure MLS.

credentials

error
OpenMLS Errors

extensions

framing

group
Group APIs for MLS

key_packages

messages

prelude
Single place, re-exporting the most used public functions. Prelude for OpenMLS. Include this to get access to all the public functions of OpenMLS.

schedule
This module represents the key schedule as introduced in Section 8 of the MLS specification. The key schedule evolves in epochs, where in each epoch new key material is injected.

tree

utils