Module openmls::ciphersuite

Ciphersuites for MLS

This file contains the API to interact with ciphersuites. See codec.rs and ciphersuites.rs for internals.

Modules

ciphersuites

codec
Codec implementations for the ciphersuites. Provides encoding and decoding functionality.

errors
Crypto errors

ser

signable

Structs

AeadKey

AeadNonce

Ciphersuite

HpkeCiphertext
7.7. Update Paths

KdfLabel
KdfLabel is later serialized and used in the label field of kdf_expand_label.

Mac
9.2 Message framing

ReuseGuard

Secret
A struct to contain secrets. This is to provide better visibility into where and how secrets are used and to avoid passing secrets in their raw representation.

Signature

SignatureKeypair

SignaturePrivateKey

SignaturePublicKey

Enums

CiphersuiteName

SignatureScheme
SignatureScheme according to IANA TLS parameters

Constants

NONCE_BYTES

REUSE_GUARD_BYTES

Functions

equal_ct
Compare two byte slices in a way that’s hopefully not optimised out by the compiler.