Trait openmls::ciphersuite::signable::Signable

pub trait Signable: Sized {
    type SignedOutput;
    fn unsigned_payload(&self) -> Result<Vec<u8>, CodecError>;

    fn sign(
        self, 
        credential_bundle: &CredentialBundle
    ) -> Result<Self::SignedOutput, CredentialError>
    where
        Self::SignedOutput: SignedStruct<Self>,
    { ... }
}

The Signable trait is implemented by all struct that are being signed. The implementation has to provide the unsigned_payload function.

Associated Types

type SignedOutput

Loading content...
Required methods

fn unsigned_payload(&self) -> Result<Vec<u8>, CodecError>

Return the unsigned, serialized payload that should be signed.

Loading content...
Provided methods

fn sign(
self,
credential_bundle: &CredentialBundle
) -> Result<Self::SignedOutput, CredentialError> where
Self::SignedOutput: SignedStruct<Self>,

Sign the payload with the given id.

Returns a Signature.

Loading content...