Trait openmls::ciphersuite::signable::Signable

pub trait Signable: Sized {
    type SignedOutput;

    fn unsigned_payload(&self) -> Result<Vec<u8>, Error>;
    fn label(&self) -> &str;

    fn sign(
        self,
        backend: &impl OpenMlsCryptoProvider,
        credential_bundle: &CredentialBundle
    ) -> Result<Self::SignedOutput, LibraryError>
    where
        Self::SignedOutput: SignedStruct<Self>,
    { ... }
}

The Signable trait is implemented by all struct that are being signed. The implementation has to provide the unsigned_payload function.

Required Associated Types

type SignedOutput

The type of the object once it’s signed.

Required Methods

fn unsigned_payload(&self) -> Result<Vec<u8>, Error>

Return the unsigned, serialized payload that should be signed.

fn label(&self) -> &str

Return the string label used for labeled signing.

Provided Methods

fn sign(
self,
backend: &impl OpenMlsCryptoProvider,
credential_bundle: &CredentialBundle
) -> Result<Self::SignedOutput, LibraryError>where
Self::SignedOutput: SignedStruct<Self>,

Sign the payload with the given id.

Returns a Signature.