Trait openmls::ciphersuite::signable::Signable

pub trait Signable: Sized {
    fn unsigned_payload(&self) -> Result<Vec<u8>, CodecError>;

    fn sign(&self, credential_bundle: &CredentialBundle) -> Signature { ... }

    fn verify(&self, credential: &Credential, signature: &Signature) -> bool { ... }
}

The Signable trait is implemented by all struct that are being signed. The implementation has to provide the unsigned_payload function.

Required methods

fn unsigned_payload(&self) -> Result<Vec<u8>, CodecError>

Loading content...
Provided methods

fn sign(&self, credential_bundle: &CredentialBundle) -> Signature

Sign the payload with the given id.

Returns a Signature.

fn verify(&self, credential: &Credential, signature: &Signature) -> bool

Verifies the payload against the given credential and signature.

Returns a true if the signature is valid and false otherwise.

Loading content...