openmls::ciphersuite::signable
Trait SignedStruct

pub trait SignedStruct<T> {
    // Required method
    fn from_payload(payload: T, signature: Signature) -> Self;
}

This trait must be implemented by all structs that contain a self-signature.

Required Methods

fn from_payload(payload: T, signature: Signature) -> Self

Build a signed struct version from the payload struct.

Object Safety

This trait is not object safe.