openmls::ciphersuite::signable

Trait SignedStruct

pub trait SignedStruct<T> {
    // Required method
    fn from_payload(
        payload: T,
        signature: Signature,
        serialized_payload: Vec<u8>,
    ) -> Self;
}

Available on 32-bit or 64-bit only.

This trait must be implemented by all structs that contain a self-signature.

Required Methods

fn from_payload( payload: T, signature: Signature, serialized_payload: Vec<u8>, ) -> Self

Build a signed struct version from the payload struct.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.