Trait openmls::ciphersuite::signable::SignedStruct

pub trait SignedStruct<T> {
    fn from_payload(payload: T, signature: Signature) -> Self;
}

This trait must be implemented by all structs that contain a self-signature.

Required methods

fn from_payload(payload: T, signature: Signature) -> Self

Build a signed struct version from the payload struct.

Loading content...