Trait openmls::ciphersuite::signable::VerifiedStruct

pub trait VerifiedStruct<T> {
    type SealingType: Default;
}

This trait must be implemented by all structs that contain a verified self-signature.

Associated Types

type SealingType: Default

This type is used to prevent users of the trait from bypassing verify by simply calling from_verifiable. Seal should be a dummy type defined in a private module as follows:

mod private_mod {
    pub struct Seal;

    impl Default for Seal {
        fn default() -> Self {
            Seal {}
        }
    }
}

Loading content...