Trait openmls::ciphersuite::signable::VerifiedStruct

pub trait VerifiedStruct<T> {
    fn from_verifiable(verifiable: T) -> Self;
}

This trait must be implemented by all structs that contain a verified self-signature.

Required methods

fn from_verifiable(verifiable: T) -> Self

Build a verified struct version from the payload struct.

Loading content...