openmls

Module credentials

Credentials

A Credential contains identifying information about the client that created it. Credentials represent clients in MLS groups and are used to authenticate their messages. Each KeyPackage, as well as each client (leaf node) in the group (tree), contains a Credential and is authenticated.

The Credential must be checked by an authentication server and the application. This process is out of scope for MLS.

Clients can create a Credential.

The MLS protocol allows the Credential representing a client in a group to change over time. Concretely, members can issue an Update proposal or a Full Commit to update their LeafNode, including the Credential in it. The Update must be authenticated using the signature public key corresponding to the old Credential.

When receiving a credential update from another member, applications must query the Authentication Service to ensure the new credential is valid.

There are multiple CredentialTypes, although OpenMLS currently only supports the BasicCredential.

Modules

errors

Credential errors

Structs

BasicCredential

Basic Credential.

Certificate

X.509 Certificate.

Credential

Credential.

CredentialWithKey

A wrapper around a credential with a corresponding public key.

NewSignerBundle

Bundle consisting of a [Signer] and a CredentialWithKey to be used to update the signature key in an MlsGroup. The public key and credential in credential_with_key MUST match the signature key exposed by signer.

Enums

CredentialType

CredentialType.