openmls
Module extensions

Extensions

In MLS, extensions appear in the following places:

In KeyPackages, to describe client capabilities and aspects of their participation in the group.

In the GroupInfo, to tell new members of a group what parameters are being used by the group, and to provide any additional details required to join the group.

In the GroupContext object, to ensure that all members of the group have the same view of the parameters in use.

Note that GroupInfo and GroupContext are not exposed in OpenMLS’ public API.

OpenMLS supports the following extensions:

ApplicationIdExtension (KeyPackage extension)

RatchetTreeExtension (GroupInfo extension)

RequiredCapabilitiesExtension (GroupContext extension)

ExternalPubExtension (GroupInfo extension)

Modules

errors

Extension errors.

Structs

ApplicationIdExtension

Application Identifiers

Extensions

A list of extensions with unique extension types.

ExternalPubExtension

ExternalSender

ExternalSender

LastResortExtension

RatchetTreeExtension

Ratchet Tree Extension.

RequiredCapabilitiesExtension

Required Capabilities Extension.

SenderExtensionIndex

Identifies an external sender in the ExternalSendersExtension.

UnknownExtension

A unknown/unparsed extension represented by raw bytes.

Enums

Extension

Extension

ExtensionType

MLS Extension Types

Type Aliases

ExternalSendersExtension

ExternalSender (extension data)