openmls

Module extensions

Available on 32-bit or 64-bit only.

Extensions

In MLS, extensions appear in the following places:

In KeyPackages, to describe client capabilities and aspects of their participation in the group.

In GroupInfo, to inform new members of the group’s parameters and to provide any additional information required to join the group.

In the GroupContext object, to ensure that all members of the group have a consistent view of the parameters in use.

Note that GroupInfo and GroupContext are not exposed via OpenMLS’ public API.

OpenMLS supports the following extensions:

ApplicationIdExtension (KeyPackage extension)

RatchetTreeExtension (GroupInfo extension)

RequiredCapabilitiesExtension (GroupContext extension)

ExternalPubExtension (GroupInfo extension)

Modules

errors

Extension errors.

Structs

AnyObject

Any object

AppDataDictionaryextensions-draft-08

Serializable app data dictionary in the AppDataDictionaryExtension.

AppDataDictionaryExtensionextensions-draft-08

App Data Dictionary Extension.

ApplicationIdExtension

Application Identifiers

Extensions

A Extension for Object of type T

ExternalPubExtension

ExternalSender

ExternalSender

LastResortExtension

RatchetTreeExtension

Ratchet Tree Extension.

RequiredCapabilitiesExtension

Required Capabilities Extension.

SenderExtensionIndex

Identifies an external sender in the ExternalSendersExtension.

UnknownExtension

A unknown/unparsed extension represented by raw bytes.

Enums

Extension

Extension

ExtensionType

MLS Extension Types

Traits

ExtensionValidator

Can be implemented by a type to validate extensions.

Type Aliases

ExternalSendersExtension

ExternalSender (extension data)