Module openmls::extensions

Extensions

Extension struct

An extension has an ExtensionType and an opaque payload (byte vector). This isn’t used in OpenMLS at all but part of the (de)serialization process of each extension.

See IANA registry for registered values

uint16 ExtensionType;

struct {
    ExtensionType extension_type;
    opaque extension_data<0..2^32-1>;
} Extension;

Re-exports

pub use errors::*;

Modules

errors

Extension errors.

Structs

CapabilitiesExtension

KeyIdExtension

LifetimeExtension

The lifetime extension holds a not before and a not after time measured in seconds since the Unix epoch (1970-01-01T00:00:00Z).

ParentHashExtension

RatchetTreeExtension

RequiredCapabilitiesExtension

Enums

Extension

An extension can be one of the following elements.

ExtensionType

Extension types