Module openmls::extensions

Re-exports

pub use capabilities_extension::CapabilitiesExtension;

pub use errors::*;

pub use key_package_id_extension::KeyIDExtension;

pub use life_time_extension::LifetimeExtension;

Modules

capabilities_extension
Capabilities Extension

errors
Extension errors.

key_package_id_extension
KeyPackage Identifiers

life_time_extension
Life time extension

parent_hash_extension
Parent hash extension

ratchet_tree_extension
Ratchet tree extension

Structs

ExtensionStruct
Extension struct

Enums

ExtensionType
Extension types

Constants

_Extension_registry

Traits

Extension
Extension

ExtensionHelper

Functions

encode_extensions
Encode extensions with TLS encoding. This is used whenever a struct contains extensions.

extensions_vec_from_cursor
Read a list of extensions from a Cursor into a vector of Extensions.

from_bytes
Build a new extension of the given type from a byte slice.