Module openmls::extensions

Extensions

In MLS, extensions appear in the following places:

In KeyPackages, to describe client capabilities and aspects of their participation in the group.

In the GroupInfo, to tell new members of a group what parameters are being used by the group, and to provide any additional details required to join the group.

In the GroupContext object, to ensure that all members of the group have the same view of the parameters in use.

Note that GroupInfo and GroupContext are not exposed in OpenMLS’ public API.

OpenMLS supports the following extensions:

ApplicationIdExtension (KeyPackage extension)

RatchetTreeExtension (GroupInfo extension)

RequiredCapabilitiesExtension (GroupContext extension)

ExternalPubExtension (GroupInfo extension)

CapabilitiesExtension (KeyPackage extension)

LifetimeExtension (KeyPackage extension)

ParentHashExtension (KeyPackage extension)

Modules

errors

Extension errors.

Structs

ApplicationIdExtension

Application Identifiers

CapabilitiesExtension

Capabilities Extension

ExternalPubExtension

LifetimeExtension

Life time extension

ParentHashExtension

Parent hash extension

RatchetTreeExtension

Ratchet Tree Extension.

RequiredCapabilitiesExtension

Required Capabilities Extension.

Enums

Extension

Extension

ExtensionType

MLS Extension Types

Type Definitions

ExternalSendersExtension

ExternalSender (extension data)