Module openmls::extensions::life_time_extension

Life time extension

KeyPackage Extension

7.2. Lifetime

The lifetime extension represents the times between which clients will consider a KeyPackage valid. This time is represented as an absolute time, measured in seconds since the Unix epoch (1970-01-01T00:00:00Z). A client MUST NOT use the data in a KeyPackage for any processing before the not_before date, or after the not_after date.

Applications MUST define a maximum total lifetime that is acceptable for a KeyPackage, and reject any KeyPackage where the total lifetime is longer than this duration.This extension MUST always be present in a KeyPackage.

uint64 not_before;
uint64 not_after;

Structs

LifetimeExtension
The lifetime extension holds a not before and a not after time measured in seconds since the Unix epoch (1970-01-01T00:00:00Z).

Statics

__init6011765713567273734___rust_ctor___ctor

Functions

__init6011765713567273734