openmls::framing

Struct ProcessedMessage

pub struct ProcessedMessage { /* private fields */ }

A message that has passed all syntax and semantics checks.

Implementations

impl ProcessedMessage

pub fn group_id(&self) -> &GroupId

Returns the group ID of the message.

pub fn epoch(&self) -> GroupEpoch

Returns the epoch of the message.

pub fn sender(&self) -> &Sender

Returns the sender of the message.

pub fn aad(&self) -> &[u8]

Returns the additional authenticated data (AAD) of the message.

pub fn content(&self) -> &ProcessedMessageContent

Returns the content of the message.

pub fn into_content(self) -> ProcessedMessageContent

Returns the content of the message and consumes the message.

pub fn credential(&self) -> &Credential

Returns the credential of the message.