Module openmls::group

Group API for MLS

This module contains the API to interact with groups.

Modules

errors

MLS group errors

mls_group

MLS Group

Structs

GroupEpoch

Group epoch. Internally this is stored as a u64. The group epoch is incremented with every valid Commit that is merged into the group state.

GroupId

A group ID. The group ID is chosen by the creator of the group and should be globally unique.

ProposalStore

A ProposalStore can store the standalone proposals that are received from the DS in between two commit messages.

QueuedAddProposal

A queued Add proposal

QueuedProposal

Alternative representation of a Proposal, where the sender is extracted from the encapsulating MlsPlaintext and the ProposalRef is attached.

QueuedPskProposal

A queued PresharedKey proposal

QueuedRemoveProposal

A queued Remove proposal

QueuedUpdateProposal

A queued Update proposal

StagedCommit

Contains the changes from a commit to the group state.