Module openmls::group

Group APIs for MLS

This file contains the API to interact with groups.

The low-level standard API is described in the Api trait.
The high-level API is exposed in ManagedGroup.

Re-exports

pub use errors::ApplyCommitError;

pub use errors::CreateCommitError;

pub use errors::ExporterError;

pub use errors::GroupError;

pub use errors::WelcomeError;

Modules

callbacks

codec

config

errors
MLS Group errors

errors
MLS Managed Group errors

group_context

managed_group

mls_group

Structs

GroupConfig
Configuration for an MLS group.

GroupContext

GroupEpoch

GroupId

ManagedGroup
A ManagedGroup represents an MlsGroup with an easier, high-level API designed to be used in production. The API exposes high level functions to manage a group by adding/removing members, get the current member list, etc.

ManagedGroupCallbacks
Collection of callback functions that are passed to a ManagedGroup as part of the configurations. All callback functions are optional.

ManagedGroupConfig
Specifies the configuration parameters for a managed group

MlsGroup

UpdatePolicy
Specifies in which intervals the own leaf node should be updated

Enums

EmptyInputError

HandshakeMessageFormat
Defines whether handshake messages (Proposals & Commits) are encrypted. Application are always encrypted regardless. Plaintext: Handshake messages are returned as MLSPlaintext messages Ciphertext: Handshake messages are returned as MLSCiphertext messages

InvalidMessageError

MLSMessage
Unified message type

ManagedGroupError

PendingProposalsError

Removal
This enum lists the 4 different variants of a removal, depending on who the remover and who the leaver is.

UseAfterEviction

Type Definitions

AppMessageReceived

AutoSave

CreateCommitResult

ErrorOccurred

InvalidMessageReceived

MemberAdded

MemberRemoved

MemberUpdated

PskFetcher
This callback function is used in several places in MlsGroup. It gets called whenever the key schedule is advanced and references to PSKs are encountered. Since the PSKs are to be trandmitted out-of-band, they need to be fetched from wherever they are stored.

PskReceived

ReInitReceived

ValidateAdd

ValidateRemove