Module openmls::group

Group APIs for MLS

This file contains the API to interact with groups.

The low-level standard API is described in the Api trait.
The high-level API is exposed in MlsGroup.

Modules

config

errors

MLS MlsGroup errors

errors

MLS CoreGroup errors

processing

Structs

GroupEpoch

GroupId

MlsGroup

A MlsGroup represents an [CoreGroup] with an easier, high-level API designed to be used in production. The API exposes high level functions to manage a group by adding/removing members, get the current member list, etc.

MlsGroupConfig

Specifies the configuration parameters for a MlsGroup

MlsGroupConfigBuilder

Enums

EmptyInputError

InnerState

Enum that indicates whether the inner group state has been modified since the last time it was persisted. InnerState::Changed indicates that the state has changed and that [.save()] should be called. InnerState::Persisted indicates that the state has not been modified and therefore doesn’t need to be persisted.

InvalidMessageError

MlsGroupError

MlsGroupState

MlsGroupState determines the state of an MlsGroup. The different states and their transitions are as follows:

PendingCommitState

PendingProposalsError

UseAfterEviction