Module openmls::group::config

Configuration module for MlsGroup configurations.

Structs

ALL_VALID_WIRE_FORMAT_POLICIES

All valid wire format policy combinations

MIXED_CIPHERTEXT_WIRE_FORMAT_POLICY

Mixed ciphertext wire format policy combination.

MIXED_PLAINTEXT_WIRE_FORMAT_POLICY

Mixed plaintext wire format policy combination.

MlsGroupConfig

Specifies the configuration parameters for a MlsGroup. Refer to the User Manual for more information about the different configuration values.

MlsGroupConfigBuilder

Builder for an MlsGroupConfig.

PURE_CIPHERTEXT_WIRE_FORMAT_POLICY

Pure ciphertext wire format policy.

PURE_PLAINTEXT_WIRE_FORMAT_POLICY

Pure plaintext wire format policy.

WireFormatPolicy

Defines what wire format is desired for outgoing handshake messages. Note that application messages must always be encrypted.

Enums

IncomingWireFormatPolicy

Defines what wire format is acceptable for incoming handshake messages. Note that application messages must always be encrypted.

OutgoingWireFormatPolicy

Defines what wire format should be used for outgoing handshake messages. Note that application messages must always be encrypted.