Module openmls::group::errors

MLS group errors

This module contains errors that originate at lower levels and are partially re-exported in errors thrown by functions of the MlsGroup API.

Enums

AddMembersError

Add members error

CommitToPendingProposalsError

Commit to pending proposals error

CreateCommitError

Create commit error

CreateMessageError

Create message error

EmptyInputError

EmptyInput error

ExportGroupInfoError

Errors that can happen when exporting a group info object.

ExportSecretError

Export secret error

ExternalCommitError

External Commit error

ExternalCommitValidationError

External Commit validaton error

LeaveGroupError

Leave group error

MlsGroupStateError

Group state error

NewGroupError

New group error

ProcessMessageError

Process message error

ProposalValidationError

Proposal validation error

ProposeAddMemberError

Propose add members error

ProposeRemoveMemberError

Propose remove members error

ProposeSelfUpdateError

Propose self update error

RemoveMembersError

Remove members error

SelfUpdateError

Self update error

StageCommitError

Stage Commit error

ValidationError

Validation error

WelcomeError

Welcome error