Module openmls::group::mls_group

Modules

apply_commit

create_commit

new_from_welcome

Structs

MlsGroup

Functions

psk_output

update_confirmed_transcript_hash

update_interim_transcript_hash

Type Definitions

CreateCommitResult

PskFetcher
This callback function is used in several places in MlsGroup. It gets called whenever the key schedule is advanced and references to PSKs are encountered. Since the PSKs are to be trandmitted out-of-band, they need to be fetched from wherever they are stored.