Struct openmls::group::mls_group::MlsGroup

pub struct MlsGroup { /* fields omitted */ }

Implementations

impl MlsGroup

pub fn create_commit(
&self,
params: CreateCommitParams<'_>,
backend: &impl OpenMlsCryptoProvider
) -> CreateCommitResult

impl MlsGroup

pub fn parse_message(
&mut self,
message: MlsMessageIn,
backend: &impl OpenMlsCryptoProvider
) -> Result<UnverifiedMessage, MlsGroupError>

This function is used to parse messages from the DS. It checks for syntactic errors and makes some semantic checks as well. If the input is a MlsCiphertext message, it will be decrypted. Returns an UnverifiedMessage that can be inspected and later processed in Self::process_unverified_message(). Checks the following semantic validation:

ValSem2

ValSem3

ValSem4

ValSem5

ValSem6

ValSem7

ValSem9

pub fn process_unverified_message(
&mut self,
unverified_message: UnverifiedMessage,
signature_key: Option<&SignaturePublicKey>,
proposal_store: &ProposalStore,
own_kpbs: &[KeyPackageBundle],
backend: &impl OpenMlsCryptoProvider
) -> Result<ProcessedMessage, MlsGroupError>

This processing function does most of the semantic verifications. It returns a ProcessedMessage enum. Checks the following semantic validation:

ValSem8

ValSem10

ValSem100

ValSem101

ValSem102

ValSem103

ValSem104

ValSem105

ValSem106

ValSem107

ValSem109

ValSem110

pub fn merge_staged_commit(
&mut self,
staged_commit: StagedCommit,
proposal_store: &mut ProposalStore
) -> Result<(), MlsGroupError>

Merge a StagedCommit into the group after inspection

impl MlsGroup

pub fn stage_commit(
&mut self,
mls_plaintext: &MlsPlaintext,
proposal_store: &ProposalStore,
own_key_packages: &[KeyPackageBundle],
psk_fetcher_option: Option<PskFetcher>,
backend: &impl OpenMlsCryptoProvider
) -> Result<StagedCommit, MlsGroupError>

Stages a commit message. This function does the following:

Applies the proposals covered by the commit to the tree

Applies the (optional) update path to the tree

Calculates the path secrets

Initializes the key schedule for epoch rollover

Verifies the confirmation tag/membership tag Returns a StagedCommit that can be inspected and later merged into the group state with MlsGroup::merge_commit() This function does the following checks:

ValSem100

ValSem101

ValSem102

ValSem103

ValSem104

ValSem105

ValSem106

ValSem107

ValSem108

ValSem109

ValSem110

ValSem201

ValSem205

pub fn merge_commit(
&mut self,
staged_commit: StagedCommit
) -> Result<(), MlsGroupError>

Merges a StagedCommit into the group state.

This function should not fail and only returns a Result, because it might throw a LibraryError.

impl MlsGroup

pub fn validate_framing(
&self,
message: &MlsMessageIn
) -> Result<(), MlsGroupError>

Checks the following semantic validation:

ValSem2

ValSem3

pub fn validate_plaintext(
&self,
plaintext: &VerifiableMlsPlaintext
) -> Result<(), MlsGroupError>

Checks the following semantic validation:

ValSem4

ValSem5

ValSem7

ValSem9

pub fn validate_add_proposals(
&self,
staged_proposal_queue: &StagedProposalQueue
) -> Result<(), MlsGroupError>

Validate Add proposals. This function implements the following checks:

ValSem100

ValSem101

ValSem102

ValSem103

ValSem104

ValSem105

TODO: ValSem106

pub fn validate_remove_proposals(
&self,
staged_proposal_queue: &StagedProposalQueue
) -> Result<(), MlsGroupError>

Validate Remove proposals. This function implements the following checks:

ValSem107

ValSem108

pub fn validate_update_proposals(
&self,
staged_proposal_queue: &StagedProposalQueue,
path_key_package: Option<(LeafIndex, &KeyPackage)>
) -> Result<(), MlsGroupError>

Validate Update proposals. This function implements the following checks:

ValSem109

ValSem110

impl MlsGroup

Public MlsGroup functions.

pub fn builder(
group_id: GroupId,
key_package_bundle: KeyPackageBundle
) -> MlsGroupBuilder

Get a builder for MlsGroup.

pub fn new_from_welcome(
welcome: Welcome,
nodes_option: Option<Vec<Option<Node>>>,
kpb: KeyPackageBundle,
psk_fetcher_option: Option<PskFetcher>,
backend: &impl OpenMlsCryptoProvider
) -> Result<Self, MlsGroupError>

pub fn create_add_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
joiner_key_package: KeyPackage,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

pub fn create_update_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
key_package: KeyPackage,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

pub fn create_remove_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
removed_index: LeafIndex,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

pub fn create_presharedkey_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
psk: PreSharedKeyId,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

pub fn create_group_context_ext_proposal(
&self,
framing_parameters: FramingParameters<'_>,
credential_bundle: &CredentialBundle,
extensions: &[Extension],
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

Create a GroupContextExtensions proposal.

pub fn create_application_message(
&mut self,
aad: &[u8],
msg: &[u8],
credential_bundle: &CredentialBundle,
padding_size: usize,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsCiphertext, MlsGroupError>

pub fn encrypt(
&mut self,
mls_plaintext: MlsPlaintext,
padding_size: usize,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsCiphertext, MlsGroupError>

pub fn decrypt(
&mut self,
mls_ciphertext: &MlsCiphertext,
backend: &impl OpenMlsCryptoProvider
) -> Result<VerifiableMlsPlaintext, MlsGroupError>

Decrypt an MlsCiphertext into an MlsPlaintext

pub fn verify(
&self,
verifiable: VerifiableMlsPlaintext,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsPlaintext, MlsGroupError>

Set the context of the VerifiableMlsPlaintext (if it has not been set already), verify it and return the MlsPlaintext.

pub fn verify_membership_tag(
&self,
backend: &impl OpenMlsCryptoProvider,
verifiable_mls_plaintext: &mut VerifiableMlsPlaintext
) -> Result<(), MlsGroupError>

Set the context of the UnverifiedMlsPlaintext and verify its membership tag.

pub fn export_secret(
&self,
backend: &impl OpenMlsCryptoProvider,
label: &str,
context: &[u8],
key_length: usize
) -> Result<Vec<u8>, MlsGroupError>

Exporter

pub fn authentication_secret(&self) -> Vec<u8>

Returns the authentication secret

pub fn load<R: Read>(reader: R) -> Result<MlsGroup, Error>

Loads the state from persisted state

pub fn save<W: Write>(&self, writer: &mut W) -> Result<(), Error>

Persists the state

pub fn treesync(&self) -> &TreeSync

Returns the ratchet tree

pub fn ciphersuite(&self) -> &'static Ciphersuite

Get the ciphersuite implementation used in this group.

pub fn context(&self) -> &GroupContext

Get the group context

pub fn group_id(&self) -> &GroupId

Get the group ID

pub fn members(&self) -> Result<BTreeMap<LeafIndex, &Credential>, MlsGroupError>

Get the members of the group, indexed by their leaves.

pub fn other_extensions(&self) -> Vec<Extension>

Get the groups extensions. Right now this is limited to the ratchet tree extension which is built on the fly when calling this function.

pub fn group_context_extensions(&self) -> &[Extension]

Get the group context extensions.

pub fn required_capabilities(&self) -> Option<&RequiredCapabilitiesExtension>

Get the required capabilities extension of this group.

pub fn export_public_group_state(
&self,
backend: &impl OpenMlsCryptoProvider,
credential_bundle: &CredentialBundle
) -> Result<PublicGroupState, CredentialError>

Export the PublicGroupState

pub fn use_ratchet_tree_extension(&self) -> bool

Returns true if the group uses the ratchet tree extension anf `false otherwise