Struct openmls::group::MlsGroup

pub struct MlsGroup { /* fields omitted */ }

Implementations

impl MlsGroup

Public MlsGroup functions.

pub fn new(
id: &[u8],
ciphersuite_name: CiphersuiteName,
key_package_bundle: KeyPackageBundle,
config: GroupConfig,
psk_option: impl Into<Option<PskSecret>>,
version: impl Into<Option<ProtocolVersion>>
) -> Result<Self, GroupError>

pub fn new_from_welcome(
welcome: Welcome,
nodes_option: Option<Vec<Option<Node>>>,
kpb: KeyPackageBundle,
psk_fetcher_option: Option<PskFetcher>
) -> Result<Self, GroupError>

pub fn create_add_proposal(
&self,
aad: &[u8],
credential_bundle: &CredentialBundle,
joiner_key_package: KeyPackage
) -> Result<MlsPlaintext, GroupError>

pub fn create_update_proposal(
&self,
aad: &[u8],
credential_bundle: &CredentialBundle,
key_package: KeyPackage
) -> Result<MlsPlaintext, GroupError>

pub fn create_remove_proposal(
&self,
aad: &[u8],
credential_bundle: &CredentialBundle,
removed_index: LeafIndex
) -> Result<MlsPlaintext, GroupError>

pub fn create_presharedkey_proposal(
&self,
aad: &[u8],
credential_bundle: &CredentialBundle,
psk: PreSharedKeyId
) -> Result<MlsPlaintext, GroupError>

pub fn create_commit(
&self,
aad: &[u8],
credential_bundle: &CredentialBundle,
proposals_by_reference: &[&MlsPlaintext],
proposals_by_value: &[&Proposal],
force_self_update: bool,
psk_fetcher_option: Option<PskFetcher>
) -> CreateCommitResult

pub fn apply_commit(
&mut self,
mls_plaintext: &MlsPlaintext,
proposals: &[&MlsPlaintext],
own_key_packages: &[KeyPackageBundle],
psk_fetcher_option: Option<PskFetcher>
) -> Result<(), GroupError>

pub fn create_application_message(
&mut self,
aad: &[u8],
msg: &[u8],
credential_bundle: &CredentialBundle,
padding_size: usize
) -> Result<MlsCiphertext, GroupError>

pub fn encrypt(
&mut self,
mls_plaintext: MlsPlaintext,
padding_size: usize
) -> Result<MlsCiphertext, GroupError>

pub fn decrypt(
&mut self,
mls_ciphertext: &MlsCiphertext
) -> Result<MlsPlaintext, MlsCiphertextError>

Decrypt an MlsCiphertext into an MlsPlaintext

pub fn verify(
&self,
verifiable: VerifiableMlsPlaintext<'_>
) -> Result<MlsPlaintext, MlsCiphertextError>

Verify a VerifiableMlsPlaintext and get the MlsPlaintext.

pub fn verify_membership_tag(
&self,
mls_plaintext: &MlsPlaintext
) -> Result<(), MlsPlaintextError>

Verify the membership tag an MlsPlaintext

pub fn export_secret(
&self,
label: &str,
context: &[u8],
key_length: usize
) -> Result<Vec<u8>, GroupError>

Exporter

pub fn authentication_secret(&self) -> Vec<u8>

Returns the authentication secret

pub fn load<R: Read>(reader: R) -> Result<MlsGroup, Error>

Loads the state from persisted state

pub fn save<W: Write>(&self, writer: &mut W) -> Result<(), Error>

Persists the state

pub fn tree(&self) -> Ref<'_, RatchetTree>

Returns the ratchet tree

pub fn ciphersuite(&self) -> &'static Ciphersuite

Get the ciphersuite implementation used in this group.

pub fn context(&self) -> &GroupContext

Get the group context

pub fn group_id(&self) -> &GroupId

Get the group ID

pub fn extensions(&self) -> Vec<Box<dyn Extension>>

Get the groups extensions. Right now this is limited to the ratchet tree extension which is built on the fly when calling this function.

pub fn export_public_group_state(
&self,
credential_bundle: &CredentialBundle
) -> Result<PublicGroupState, CredentialError>

Export the PublicGroupState

pub fn use_ratchet_tree_extension(&self) -> bool

Returns true if the group uses the ratchet tree extension anf `false otherwise