Struct openmls::group::MlsGroup

pub struct MlsGroup { /* private fields */ }

A MlsGroup represents an MLS group with a high-level API. The API exposes high level functions to manage a group by adding/removing members, get the current member list, etc.

The API is modeled such that it can serve as a direct interface to the Delivery Service. Functions that modify the public state of the group will return a Vec<MLSMessageOut> that can be sent to the Delivery Service directly. Conversely, incoming messages from the Delivery Service can be fed into process_message().

An MlsGroup has an internal queue of pending proposals that builds up as new messages are processed. When creating proposals, those messages are not automatically appended to this queue, instead they have to be processed again through process_message(). This allows the Delivery Service to reject them (e.g. if they reference the wrong epoch).

If incoming messages or applied operations are semantically or syntactically incorrect, an error event will be returned with a corresponding error message and the state of the group will remain unchanged.

An MlsGroup has an internal state variable determining if it is active or inactive, as well as if it has a pending commit. See MlsGroupState for more information.

Implementations

impl MlsGroup

pub fn create_message(
&mut self,
backend: &impl OpenMlsCryptoProvider,
message: &[u8]
) -> Result<MlsMessageOut, CreateMessageError>

Creates an application message. Returns CreateMessageError::MlsGroupStateError::UseAfterEviction if the member is no longer part of the group. Returns CreateMessageError::MlsGroupStateError::PendingProposal if pending proposals exist. In that case .process_pending_proposals() must be called first and incoming messages from the DS must be processed afterwards.

impl MlsGroup

pub fn new(
backend: &impl OpenMlsCryptoProvider,
mls_group_config: &MlsGroupConfig,
key_package_hash: &[u8]
) -> Result<Self, NewGroupError>

Creates a new group with the creator as the only member (and a random group ID).

This function removes the KeyPackageBundle corresponding to the key_package_hash from the key store. Returns an error (NewGroupError::NoMatchingKeyPackageBundle) if no KeyPackageBundle can be found.

pub fn new_with_group_id(
backend: &impl OpenMlsCryptoProvider,
mls_group_config: &MlsGroupConfig,
group_id: GroupId,
key_package_hash: &[u8]
) -> Result<Self, NewGroupError>

Creates a new group with a given group ID with the creator as the only member.

This function removes the KeyPackageBundle corresponding to the key_package_hash from the key store. Returns an error (NewGroupError::NoMatchingKeyPackageBundle) if no KeyPackageBundle can be found.

pub fn new_from_welcome(
backend: &impl OpenMlsCryptoProvider,
mls_group_config: &MlsGroupConfig,
welcome: Welcome,
ratchet_tree: Option<Vec<Option<Node>>>
) -> Result<Self, WelcomeError>

Creates a new group from a Welcome message. Returns an error (WelcomeError::NoMatchingKeyPackageBundle) if no KeyPackageBundle can be found.

pub fn join_by_external_commit(
backend: &impl OpenMlsCryptoProvider,
tree_option: Option<&[Option<Node>]>,
verifiable_group_info: VerifiableGroupInfo,
mls_group_config: &MlsGroupConfig,
aad: &[u8],
credential_bundle: &CredentialBundle
) -> Result<(Self, MlsMessageOut), ExternalCommitError>

Join an existing group through an External Commit. The resulting MlsGroup instance starts off with a pending commit (the external commit, which adds this client to the group). Merging this commit is necessary for this MlsGroup instance to function properly, as, for example, this client is not yet part of the tree. As a result, it is not possible to clear the pending commit. If the external commit was rejected due to an epoch change, the MlsGroup instance has to be discarded and a new one has to be created using this function based on the latest ratchet_tree and group info. For more information on the external init process, please see Section 11.2.1 in the MLS specification.

impl MlsGroup

pub fn export_secret(
&self,
backend: &impl OpenMlsCryptoProvider,
label: &str,
context: &[u8],
key_length: usize
) -> Result<Vec<u8>, ExportSecretError>

Exports a secret from the current epoch. Returns ExportSecretError::KeyLengthTooLong if the requested key length is too long. Returns ExportSecretError::GroupStateError(MlsGroupStateError::UseAfterEviction) if the group is not active.

pub fn epoch_authenticator(&self) -> &EpochAuthenticator

Returns the epoch authenticator of the current epoch.

pub fn resumption_psk_secret(&self) -> &ResumptionPskSecret

Returns the resumption PSK secret of the current epoch.

pub fn get_past_resumption_psk(
&self,
epoch: GroupEpoch
) -> Option<&ResumptionPskSecret>

Returns a resumption psk for a given epoch. If no resumption psk is available for that epoch, None is returned.

pub fn export_group_info(
&self,
backend: &impl OpenMlsCryptoProvider,
with_ratchet_tree: bool
) -> Result<GroupInfo, ExportGroupInfoError>

Export a group info object for this group.

impl MlsGroup

pub fn self_update(
&mut self,
backend: &impl OpenMlsCryptoProvider,
key_pair: Option<HpkeKeyPair>
) -> Result<(MlsMessageOut, Option<Welcome>), SelfUpdateError>

Updates the own leaf node.

An HpkeKeyPair can optionally be provided. If not, a new one will be created on the fly.

If successful, it returns a tuple of MlsMessageOut and an optional Welcome. The Welcome is Some when the queue of pending proposals contained add proposals.

Returns an error if there is a pending commit.

pub fn propose_self_update(
&mut self,
backend: &impl OpenMlsCryptoProvider,
key_package_bundle: Option<KeyPackageBundle>
) -> Result<MlsMessageOut, ProposeSelfUpdateError>

Creates a proposal to update the own leaf node.

impl MlsGroup

pub fn add_members(
&mut self,
backend: &impl OpenMlsCryptoProvider,
key_packages: &[KeyPackage]
) -> Result<(MlsMessageOut, Welcome), AddMembersError>

Adds members to the group.

New members are added by providing a KeyPackage for each member.

This operation results in a Commit with a path, i.e. it includes an update of the committer’s leaf KeyPackage.

If successful, it returns a tuple of MlsMessageOut and Welcome.

Returns an error if there is a pending commit.

pub fn own_leaf(&self) -> Result<&LeafNode, LibraryError>

Returns a reference to the own LeafNode.

pub fn remove_members(
&mut self,
backend: &impl OpenMlsCryptoProvider,
members: &[LeafNodeIndex]
) -> Result<(MlsMessageOut, Option<Welcome>), RemoveMembersError>

Removes members from the group.

Members are removed by providing the member’s leaf index.

If successful, it returns a tuple of MlsMessageOut and an optional Welcome. The Welcome is Some when the queue of pending proposals contained add proposals

Returns an error if there is a pending commit.

pub fn propose_add_member(
&mut self,
backend: &impl OpenMlsCryptoProvider,
key_package: &KeyPackage
) -> Result<MlsMessageOut, ProposeAddMemberError>

Creates proposals to add members to the group.

Returns an error if there is a pending commit.

pub fn propose_remove_member(
&mut self,
backend: &impl OpenMlsCryptoProvider,
member: LeafNodeIndex
) -> Result<MlsMessageOut, ProposeRemoveMemberError>

Creates proposals to remove members from the group. The member has to be the member’s leaf index.

Returns an error if there is a pending commit.

pub fn leave_group(
&mut self,
backend: &impl OpenMlsCryptoProvider
) -> Result<MlsMessageOut, LeaveGroupError>

Leave the group.

Creates a Remove Proposal that needs to be covered by a Commit from a different member. The Remove Proposal is returned as a MlsMessageOut.

Returns an error if there is a pending commit.

pub fn members(&self) -> impl Iterator<Item = Member> + '_

Returns a list of Members in the group.

pub fn member(&self, leaf_index: LeafNodeIndex) -> Option<&Credential>

Returns the Credential of a member corresponding to the given leaf index. Returns None if the member can not be found in this group.

impl MlsGroup

pub fn process_message(
&mut self,
backend: &impl OpenMlsCryptoProvider,
message: MlsMessageIn
) -> Result<ProcessedMessage, ProcessMessageError>

Parses incoming messages from the DS. Checks for syntactic errors and makes some semantic checks as well. If the input is an encrypted message, it will be decrypted. This processing function does syntactic and semantic validation of the message. It returns a ProcessedMessage enum.

Errors:

Returns an ProcessMessageError when the validation checks fail with the exact reason of the failure.

pub fn store_pending_proposal(&mut self, proposal: QueuedProposal)

Stores a standalone proposal in the internal ProposalStore

pub fn commit_to_pending_proposals(
&mut self,
backend: &impl OpenMlsCryptoProvider
) -> Result<(MlsMessageOut, Option<Welcome>), CommitToPendingProposalsError>

Creates a Commit message that covers the pending proposals that are currently stored in the group’s ProposalStore.

Returns an error if there is a pending commit.

pub fn merge_staged_commit(&mut self, staged_commit: StagedCommit)

Merge a StagedCommit into the group after inspection. As this advances the epoch of the group, it also clears any pending commits.

pub fn merge_pending_commit(&mut self) -> Result<(), MlsGroupStateError>

Merges the pending StagedCommit if there is one, and clears the field by setting it to None.

impl MlsGroup

pub fn configuration(&self) -> &MlsGroupConfig

Returns the configuration.

pub fn set_configuration(&mut self, mls_group_config: &MlsGroupConfig)

Sets the configuration.

pub fn aad(&self) -> &[u8]

Returns the AAD used in the framing.

pub fn set_aad(&mut self, aad: &[u8])

Sets the AAD used in the framing.

pub fn ciphersuite(&self) -> Ciphersuite

Returns the group’s ciphersuite.

pub fn is_active(&self) -> bool

Returns whether the own client is still a member of the group or if it was already evicted

pub fn credential(&self) -> Result<&Credential, MlsGroupStateError>

Returns own credential. If the group is inactive, it returns a UseAfterEviction error.

pub fn own_identity(&self) -> Option<&[u8]>

Get the identity of the client’s Credential owning this group.

pub fn own_leaf_index(&self) -> LeafNodeIndex

Returns the leaf index of the client in the tree owning this group.

pub fn group_id(&self) -> &GroupId

Returns the group ID.

pub fn epoch(&self) -> GroupEpoch

Returns the epoch.

pub fn pending_proposals(&self) -> impl Iterator<Item = &QueuedProposal>

Returns an Iterator over pending proposals.

pub fn pending_commit(&self) -> Option<&StagedCommit>

Returns a reference to the StagedCommit of the most recently created commit. If there was no commit created in this epoch, either because this commit or another commit was merged, it returns None.

pub fn clear_pending_commit(&mut self)

Sets the group_state to MlsGroupState::Operational, thus clearing any potentially pending commits.

Note that this has no effect if the group was created through an external commit and the resulting external commit has not been merged yet. For more information, see MlsGroup::join_by_external_commit().

Use with caution! This function should only be used if it is clear that the pending commit will not be used in the group. In particular, if a pending commit is later accepted by the group, this client will lack the key material to encrypt or decrypt group messages.

pub fn load<R: Read>(reader: R) -> Result<MlsGroup, Error>

Loads the state from persisted state.

pub fn save<W: Write>(&mut self, writer: &mut W) -> Result<(), Error>

Persists the state.

pub fn state_changed(&self) -> InnerState

Returns true if the internal state has changed and needs to be persisted and false otherwise. Calling Self::save() resets the value to false.

pub fn export_ratchet_tree(&self) -> Vec<Option<Node>>

Exports the Ratchet Tree.