openmls::group

Struct ProposalQueue

pub struct ProposalQueue { /* private fields */ }

Proposal queue that helps filtering and sorting Proposals received during one epoch. The Proposals are stored in a HashMap which maps Proposal references to Proposals, such that, given a reference, a proposal can be accessed efficiently. To enable iteration over the queue in order, the ProposalQueue also contains a vector of ProposalRefs.

Implementations

impl ProposalQueue

pub fn get(&self, proposal_reference: &ProposalRef) -> Option<&QueuedProposal>

Returns proposal for a given proposal ID