Struct openmls::group::PublicGroup

pub struct PublicGroup { /* private fields */ }

This struct holds all public values of an MLS group.

Implementations

impl PublicGroup

pub fn process_message( &self, crypto: &impl OpenMlsCrypto, message: impl Into<ProtocolMessage>, ) -> Result<ProcessedMessage, ProcessMessageError>

This function is used to parse messages from the DS. It checks for syntactic errors and does semantic validation as well. It returns a ProcessedMessage enum. Checks the following semantic validation:

ValSem002

ValSem003

ValSem004

ValSem005

ValSem006

ValSem007

ValSem008

ValSem009

ValSem010

ValSem101

ValSem102

ValSem104

ValSem106

ValSem107

ValSem108

ValSem110

ValSem111

ValSem112

ValSem200

ValSem201

ValSem202: Path must be the right length

ValSem203: Path secrets must decrypt correctly

ValSem204: Public keys from Path must be verified and match the private keys from the direct path

ValSem205

ValSem240

ValSem241

ValSem242

ValSem244

ValSem245

ValSem246 (as part of ValSem010)

impl PublicGroup

pub fn merge_commit<Storage: PublicStorageProvider>( &mut self, storage: &Storage, staged_commit: StagedCommit, ) -> Result<(), MergeCommitError<Storage::PublicError>>

Merges a StagedCommit into the public group state.

impl PublicGroup

pub fn from_external<StorageProvider: PublicStorageProvider>( crypto: &impl OpenMlsCrypto, storage: &StorageProvider, ratchet_tree: RatchetTreeIn, verifiable_group_info: VerifiableGroupInfo, proposal_store: ProposalStore, ) -> Result<(Self, GroupInfo), CreationFromExternalError<StorageProvider::PublicError>>

Create a PublicGroup instance to start tracking an existing MLS group.

This function performs basic validation checks and returns an error if one of the checks fails. See CreationFromExternalError for more details.

pub fn ext_commit_sender_index( &self, commit: &StagedCommit, ) -> Result<LeafNodeIndex, LibraryError>

Returns the index of the sender of a staged, external commit.

pub fn members(&self) -> impl Iterator<Item = Member> + '_

Get an iterator over all Members of this PublicGroup.

pub fn export_ratchet_tree(&self) -> RatchetTree

Export the nodes of the public tree.

pub fn add_proposal<Storage: PublicStorageProvider>( &mut self, storage: &Storage, proposal: QueuedProposal, ) -> Result<(), Storage::PublicError>

Add the QueuedProposal to the PublicGroups internal ProposalStore.

pub fn remove_proposal<Storage: PublicStorageProvider>( &mut self, storage: &Storage, proposal_ref: &ProposalRef, ) -> Result<(), Storage::PublicError>

Remove the Proposal with the given ProposalRef from the PublicGroups internal ProposalStore.

pub fn queued_proposals<Storage: PublicStorageProvider>( &self, storage: &Storage, ) -> Result<Vec<(ProposalRef, QueuedProposal)>, Storage::PublicError>

Return all queued proposals

impl PublicGroup

pub fn ciphersuite(&self) -> Ciphersuite

Get the ciphersuite.

pub fn version(&self) -> ProtocolVersion

Get the version.

pub fn group_id(&self) -> &GroupId

Get the group id.

pub fn group_context(&self) -> &GroupContext

Get the group context.

pub fn required_capabilities(&self) -> Option<&RequiredCapabilitiesExtension>

Get the required capabilities.

pub fn confirmation_tag(&self) -> &ConfirmationTag

Get confirmation tag.

pub fn leaf(&self, leaf_index: LeafNodeIndex) -> Option<&LeafNode>

Return a reference to the leaf at the given LeafNodeIndex or None if the leaf is blank.

pub fn delete<Storage: PublicStorageProvider>( storage: &Storage, group_id: &GroupId, ) -> Result<(), Storage::PublicError>

Deletes the PublicGroup from storage.

pub fn load<Storage: PublicStorageProvider>( storage: &Storage, group_id: &GroupId, ) -> Result<Option<Self>, Storage::PublicError>

Loads the PublicGroup corresponding to a GroupId from storage.