Struct openmls::group::StagedCommit

pub struct StagedCommit { /* private fields */ }

Contains the changes from a commit to the group state.

Implementations

impl StagedCommit

pub fn add_proposals(&self) -> impl Iterator<Item = QueuedAddProposal<'_>>

Returns the Add proposals that are covered by the Commit message as in iterator over QueuedAddProposal.

pub fn remove_proposals(&self) -> impl Iterator<Item = QueuedRemoveProposal<'_>>

Returns the Remove proposals that are covered by the Commit message as in iterator over QueuedRemoveProposal.

pub fn update_proposals(&self) -> impl Iterator<Item = QueuedUpdateProposal<'_>>

Returns the Update proposals that are covered by the Commit message as in iterator over QueuedUpdateProposal.

pub fn psk_proposals(&self) -> impl Iterator<Item = QueuedPskProposal<'_>>

Returns the PresharedKey proposals that are covered by the Commit message as in iterator over QueuedPskProposal.

pub fn commit_update_key_package(&self) -> Option<&LeafNode>

Returns an optional leaf node from the Commit’s update path. A leaf node is returned for full and empty Commits, but not for partial Commits.

pub fn self_removed(&self) -> bool

Returns true if the member was removed through a proposal covered by this Commit message and false otherwise.