Struct openmls::key_packages::KeyPackageBundle

pub struct KeyPackageBundle { /* fields omitted */ }

Implementations

impl KeyPackageBundle

Public KeyPackageBundle functions.

pub fn new(
ciphersuites: &[CiphersuiteName],
credential_bundle: &CredentialBundle,
backend: &impl OpenMlsCryptoProvider,
extensions: Vec<Extension>
) -> Result<Self, KeyPackageError>

Create a new KeyPackageBundle with a fresh HpkeKeyPair. See new_with_keypair and new_with_version for details. This key package will have the default MLS version. Use new_with_version to get a key package bundle for a specific MLS version.

Returns a new KeyPackageBundle or a KeyPackageError.

pub fn new_with_version(
version: ProtocolVersion,
ciphersuites: &[CiphersuiteName],
backend: &impl OpenMlsCryptoProvider,
credential_bundle: &CredentialBundle,
extensions: Vec<Extension>
) -> Result<Self, KeyPackageError>

Create a new KeyPackageBundle with

a fresh HpkeKeyPair

the provided MLS version

the first cipher suite in the ciphersuites slice

the provided extensions See new_with_keypair for details.

Returns a new KeyPackageBundle or a KeyPackageError.

pub fn new_with_keypair(
ciphersuites: &[CiphersuiteName],
backend: &impl OpenMlsCryptoProvider,
credential_bundle: &CredentialBundle,
extensions: Vec<Extension>,
key_pair: HpkeKeyPair,
leaf_secret: Secret
) -> Result<Self, KeyPackageError>

Create a new KeyPackageBundle for the given ciphersuite, identity, and extensions, using the given HPKE key_pair.

Note that the capabilities extension gets added automatically, based on the configuration. The ciphersuite for this key package bundle is the first one in the ciphersuites list. If a capabilities extension is included in the extensions, its supported ciphersuites have to match the ciphersuites list.

Returns an DuplicateExtension error if extensions contains multiple extensions of the same type.

Returns a new KeyPackageBundle.

pub fn key_package(&self) -> &KeyPackage

Get a reference to the KeyPackage.

pub fn unsigned(self) -> KeyPackageBundlePayload

Get the unsigned payload version of this key package bundle for modificaiton.

impl KeyPackageBundle

Crate visible KeyPackageBundle functions.

pub fn leaf_secret(&self) -> &Secret

Get a reference to the leaf_secret.