Module openmls::messages

Messages

This module contains the types and implementations for Commit & Welcome messages, as well as Proposals & the public group state used for External Commits.

Modules

codec

Codec implementations for message structs.

proposals

Proposals

public_group_state

This module contains the types and implementations for the PublicGroupState and PublicGroupStateTbs structs of the MLS spec. The PublicGroupState implements type-enforced verification using the set of traits defined in the signable module in the same way as the MlsPlaintext.

Structs

ConfirmationTag

Confirmation tag field of MlsPlaintext. For type safety this is a wrapper around a Mac.

EncryptedGroupSecrets

EncryptedGroupSecrets

Welcome

Welcome message