openmls::messages

Module external_proposals

External Proposals

Contains the types and methods to build external proposal to add/remove a client from a MLS group

ReInit is not yet implemented

Structs

ExternalProposal

External Proposal where sender is External. A party outside the group can request to add or remove a member to the group. This proposal should then be committed by a group member. The sender must be pre configured within the group through the crate::extensions::ExternalSendersExtension

JoinProposal

External Add Proposal where sender is NewMemberProposal. A client outside the group can request joining the group. This proposal should then be committed by a group member. Note that this is unconstrained i.e. it works for any MLS group. This is not the case for the same external proposal with a Preconfigured sender