openmls::messages

Module proposals

Available on 32-bit or 64-bit only.

Proposals

This module defines all the different types of Proposals.

Structs

AddProposal

Add Proposal.

AppAckextensions-draft-08

AppAck object.

AppEphemeralProposalextensions-draft-08

AppEphemeral proposal.

CustomProposal

A custom proposal with semantics to be implemented by the application.

ExternalInitProposal

ExternalInit Proposal.

GroupContextExtensionProposal

GroupContextExtensions Proposal.

PreSharedKeyProposal

PreSharedKey Proposal.

ReInitProposal

ReInit Proposal.

RemoveProposal

Remove Proposal.

UpdateProposal

Update Proposal.

Enums

Proposal

Proposal.

ProposalOrRefType

11.2 Commit

ProposalType

MLS Proposal Types