openmls::messages

Module proposals

Proposals

This module defines all the different types of Proposals.

Structs

AddProposal

Add Proposal.

AppAckProposal

AppAck Proposal.

CustomProposal

A custom proposal with semantics to be implemented by the application.

ExternalInitProposal

ExternalInit Proposal.

GroupContextExtensionProposal

GroupContextExtensions Proposal.

PreSharedKeyProposal

PreSharedKey Proposal.

ReInitProposal

ReInit Proposal.

RemoveProposal

Remove Proposal.

UpdateProposal

Update Proposal.

Enums

Proposal

Proposal.

ProposalOrRefType

11.2 Commit

ProposalType

MLS Proposal Types