Module openmls::messages::proposals

Proposals

This module defines all the different types of Proposals.

To find out if a specific proposal type is supported, ProposalType::is_supported() can be used.

Structs

AddProposal

Add Proposal.

AppAckProposal

AppAck Proposal.

ExternalInitProposal

ExternalInit Proposal.

GroupContextExtensionProposal

Group Context Extensions Proposal

PreSharedKeyProposal

Preshared Key Proposal.

ReInitProposal

ReInit proposal.

RemoveProposal

Remove Proposal.

UpdateProposal

Update Proposal.

Enums

Proposal

Proposal.

ProposalType

MLS Proposal Types