Module openmls::messages::proposals

Structs

AddProposal

PreSharedKeyProposal
Preshared Key proposal 11.1.4 struct { PreSharedKeyID psk; } PreSharedKey;

ProposalQueue
Proposal queue that helps filtering and sorting the Proposals from one epoch. The Proposals are stored in a HashMap which maps Proposal references to Proposals, such that, given a reference, a proposal can be accessed efficiently. To enable iteration over the queue in order, the ProposalQueue also contains a vector of ProposalReferences.

ProposalReference
Reference to a Proposal. This can be used in Commit messages to reference proposals that have already been sent

QueuedProposal
Alternative representation of a Proposal, where the sender is extracted from the encapsulating MlsPlaintext and the ProposalReference is attached.

ReInitProposal
ReInit proposal 11.1.5 struct { opaque group_id<0..255>; ProtocolVersion version; CipherSuite cipher_suite; Extension extensions<0..2^32-1>; } ReInit;

RemoveProposal

UpdateProposal

Enums

Proposal
Proposal

ProposalOrRef
Type of Proposal, either by value or by reference

ProposalOrRefType
11.2 Commit

ProposalType