Module openmls::prelude

Single place, re-exporting the most used public functions. Prelude for OpenMLS. Include this to get access to all the public functions of OpenMLS.

Re-exports

pub use crate::group::*;

pub use crate::ciphersuite::hash_ref::KeyPackageRef;

pub use crate::ciphersuite::signable::*;

pub use crate::ciphersuite::*;

pub use crate::messages::*;

pub use crate::credentials::*;

pub use crate::versions::*;

pub use crate::extensions::*;

pub use crate::key_packages::*;

pub use crate::error::*;

Structs

ApplicationMessage

Application message received through a ProcessedMessage.

BranchPsk

Branch PSK

ExternalPsk

External PSK.

HpkeCiphertext

7.7. Update Paths

HpkeConfig

Convenience tuple struct for an HPKE configuration.

HpkeKeyPair

Helper holding a (private, public) key pair as byte vectors.

MlsMessageIn

Unified message type for incoming MLS messages.

MlsMessageOut

Unified message type for outgoing MLS messages.

PreSharedKeyId

A PreSharedKeyID is used to uniquely identify the PSKs that get injected in the key schedule.

PreSharedKeys

PreSharedKeys is a vector of PreSharedKeyIDs. struct { PreSharedKeyID psks<0..2^16-1>; } PreSharedKeys;

PskBundle

Contains the secret part of the PSK as well as the public part that is used as a marker for injection into the key schedule.

ReinitPsk

ReInit PSK.

SenderRatchetConfiguration

Stores the configuration parameters for DecryptionRatchets.

UnverifiedMessage

Partially checked and potentially decrypted message (if it was originally encrypted). Use this to inspect the Credential of the message sender and the optional aad if the original message was encrypted.

Enums

AeadType

AEAD types

ApplyUpdatePathError

Apply update path error

Ciphersuite

MLS ciphersuites.

CredentialError

An error that occurs in methods of a super::Credential.

CryptoError

Crypto errors.

Error

Trait errors.

HashType

Hash types

HpkeAeadType

AEAD Types for HPKE.

HpkeKdfType

KDF Types for HPKE

HpkeKemType

KEM Types for HPKE

Node

Container enum for leaf and parent nodes.

ProcessedMessage

A message that has passed all syntax and semantics checks.

Psk

PSK enum that can contain the different PSK types

PskType

Type of PSK.

PublicTreeError

Public tree error

Sender

All possible sender types according to the MLS protocol spec.

SignatureScheme

SignatureScheme according to IANA TLS parameters

Traits

OpenMlsCrypto

OpenMlsCryptoProvider

The OpenMLS Crypto Provider Trait

OpenMlsKeyStore

The Key Store trait

OpenMlsRand

TlsDeserializeTrait

The Deserialize trait defines functions to deserialize a byte slice to a struct or enum.

TlsSerializeTrait

The Serialize trait provides functions to serialize a struct or enum.

TlsSizeTrait

The Size trait needs to be implemented by any struct that should be efficiently serialized. This allows to collect the length of a serialized structure before allocating memory.

Type Definitions

ExporterSecret

KemOutput

LeafIndex

Index type to index the leaves in the binary tree