Module openmls::prelude

Single place, re-exporting the most used public functions. Prelude for OpenMLS. Include this to get access to all the public functions of OpenMLS.

Re-exports

pub use crate::group::GroupConfig;

pub use crate::group::errors::ApplyCommitError;

pub use crate::group::errors::CreateCommitError;

pub use crate::group::errors::GroupError;

pub use crate::group::errors::WelcomeError;

pub use crate::tree::index::LeafIndex;

pub use crate::config::*;

pub use crate::framing::errors::*;

pub use crate::framing::sender::Sender;

pub use crate::framing::*;

pub use crate::group::GroupId;

pub use crate::messages::Welcome;

Modules

errors
Extension errors.

errors
Key Package errors

Structs

AddProposal

AeadKey

AeadNonce

BasicCredential

BranchPsk
Branch PSK

CapabilitiesExtension

Certificate
Struct containing an X509 certificate chain, as per Spec.

Ciphersuite

Credential
Struct containing MLS credential data, where the data depends on the type.

CredentialBundle
This struct contains a credential and the corresponding private key.

Cursor

ExtensionStruct
Extension struct

ExternalPsk
External PSK.

ExternalPskBundle
External PSK Bundle. This contains the secret part of the PSK as well as the public part that is used as a marker for injection into the key schedule.

HpkeCiphertext
7.7. Update Paths

KeyIDExtension

KeyPackage

KeyPackageBundle

LifetimeExtension
The lifetime extension holds a not before and a not after time measured in seconds since the Unix epoch (1970-01-01T00:00:00Z).

ManagedGroup
A ManagedGroup represents an MlsGroup with an easier, high-level API designed to be used in production. The API exposes high level functions to manage a group by adding/removing members, get the current member list, etc.

ManagedGroupCallbacks
Collection of callback functions that are passed to a ManagedGroup as part of the configurations. All callback functions are optional.

ManagedGroupConfig
Specifies the configuration parameters for a managed group

MlsGroup

PreSharedKeyID
A PreSharedKeyID is used to uniquely identify the PSKs that get injected in the key schedule.

PreSharedKeyProposal
Preshared Key proposal 11.1.4 struct { PreSharedKeyID psk; } PreSharedKey;

PreSharedKeys
PreSharedKeys is a vector of PreSharedKeyIDs. struct { PreSharedKeyID psks<0..2^16-1>; } PreSharedKeys;

ReInitProposal
ReInit proposal 11.1.5 struct { opaque group_id<0..255>; ProtocolVersion version; CipherSuite cipher_suite; Extension extensions<0..2^32-1>; } ReInit;

ReinitPsk
ReInit PSK.

RemoveProposal

ReuseGuard

Secret
A struct to contain secrets. This is to provide better visibility into where and how secrets are used and to avoid passing secrets in their raw representation.

Signature

SignatureKeypair

SignaturePrivateKey

SignaturePublicKey

UpdatePolicy
Specifies in which intervals the own leaf node should be updated

UpdateProposal

Enums

CapabilitiesExtensionError

CiphersuiteName

CodecError

CredentialError

CredentialType
Enum for Credential Types. We only need this for encoding/decoding.

ExtensionError

ExtensionType
Extension types

HandshakeMessageFormat
Defines whether handshake messages (Proposals & Commits) are encrypted. Application are always encrypted regardless. Plaintext: Handshake messages are returned as MLSPlaintext messages Ciphertext: Handshake messages are returned as MLSCiphertext messages

InvalidExtensionError

InvalidMessageError

KeyPackageError

KeyPackageIdError

LifetimeExtensionError

MLSCredentialType
This enum contains the different available credentials.

MLSMessage
Unified message type

ManagedGroupError

PSKType
Type of PSK.

ParentHashError

Psk
PSK enum that can contain the different PSK types

RatchetTreeError

Removal
This enum lists the 4 different variants of a removal, depending on who the remover and who the leaver is.

SignatureScheme
SignatureScheme according to IANA TLS parameters

VecSize

Traits

Codec

Extension
Extension

ExtensionHelper

Functions

_print_tree

decode_vec

encode_vec