Struct openmls::prelude::KeyPackageBundle

pub struct KeyPackageBundle {
    pub(crate) key_package: KeyPackage,
    pub(crate) private_key: HPKEPrivateKey,
    pub(crate) leaf_secret: Secret,
}

Fields
key_package: KeyPackageprivate_key: HPKEPrivateKeyleaf_secret: Secret
Implementations

impl KeyPackageBundle

Public KeyPackageBundle functions.

pub fn new(
ciphersuites: &[CiphersuiteName],
credential_bundle: &CredentialBundle,
extensions: Vec<Box<dyn Extension>>
) -> Result<Self, KeyPackageError>

Create a new KeyPackageBundle with a fresh HPKEKeyPair. See new_with_keypair for details.

Returns a new KeyPackageBundle.

pub(crate) fn from_rekeyed_key_package(key_package: &KeyPackage) -> Self

Replace the init key in the KeyPackage with a random one and return a KeyPackageBundle with the corresponding secret values

pub fn new_with_keypair(
ciphersuites: &[CiphersuiteName],
credential_bundle: &CredentialBundle,
extensions: Vec<Box<dyn Extension>>,
key_pair: HPKEKeyPair,
leaf_secret: Secret
) -> Result<Self, KeyPackageError>

Create a new KeyPackageBundle for the given ciphersuite, identity, and extensions, using the given HPKE key_pair.

Note that the capabilities extension gets added automatically, based on the configuration. The ciphersuite for this key package bundle is the first one in the ciphersuites list. If a capabilities extension is included in the extensions, its supported ciphersuites have to match the ciphersuites list.

Returns an DuplicateExtension error if extensions contains multiple extensions of the same type.

Returns a new KeyPackageBundle.

pub fn key_package(&self) -> &KeyPackage

Get a reference to the KeyPackage.

impl KeyPackageBundle

Private KeyPackageBundle functions.

fn new_from_leaf_secret(
ciphersuites: &[CiphersuiteName],
credential_bundle: &CredentialBundle,
extensions: Vec<Box<dyn Extension>>,
leaf_secret: Secret
) -> Result<Self, KeyPackageError>

fn new_from_values(
key_package: KeyPackage,
private_key: HPKEPrivateKey,
leaf_secret: Secret
) -> Self

Assembles a new KeyPackageBundle from a KeyPackage, a HPKEPrivateKey, and a leaf secret

impl KeyPackageBundle

Crate visible KeyPackageBundle functions.

pub(crate) fn _set_private_key(&mut self, private_key: HPKEPrivateKey)

Update the private key in the bundle.

pub(crate) fn set_key_package(&mut self, key_package: KeyPackage)

Update the key package in the bundle.

pub fn key_package_mut(&mut self) -> &mut KeyPackage

Get a mutable reference to the KeyPackage.

pub(crate) fn private_key(&self) -> &HPKEPrivateKey

Get a reference to the HPKEPrivateKey.

pub(crate) fn leaf_secret(&self) -> &Secret

Get a reference to the leaf_secret.

pub(crate) fn derive_leaf_node_secret(
ciphersuite: &Ciphersuite,
leaf_secret: &Secret
) -> Secret

This function derives the leaf_node_secret from the leaf_secret as described in 5.4 Ratchet Tree Evolution

pub(crate) fn sign(&mut self, credential_bundle: &CredentialBundle)

Sign the KeyPackageBundle