Struct openmls::prelude::ManagedGroup

pub struct ManagedGroup<'a> {
    credential_bundle: &'a CredentialBundle,
    managed_group_config: ManagedGroupConfig,
    group: MlsGroup,
    pending_proposals: Vec<MLSPlaintext>,
    own_kpbs: Vec<KeyPackageBundle>,
    aad: Vec<u8>,
    resumption_secret_store: ResumptionSecretStore,
    active: bool,
}

A ManagedGroup represents an MlsGroup with an easier, high-level API designed to be used in production. The API exposes high level functions to manage a group by adding/removing members, get the current member list, etc.

The API is modeled such that it can serve as a direct interface to the Delivery Service. Functions that modify the public state of the group will return a Vec<MLSMessage> that can be sent to the Delivery Service directly. Conversely, incoming messages from the Delivery Service can be fed into process_messages().

A ManagedGroup has an internal queue of pending proposals that builds up as new messages are processed. When creating proposals, those messages are not automatically appended to this queue, instead they have to be processed again through process_messages(). This allows the Delivery Service to reject them (e.g. if they reference the wrong epoch).

If incoming messages or applied operations are semantically or syntactically incorrect, an error event will be returned with a corresponding error message and the state of the group will remain unchanged.

The application policy for the group can be enforced by implementing the validator callback functions and selectively allowing/ disallowing each operation (see ManagedGroupCallbacks)

Changes to the group state are dispatched as events through callback functions (see ManagedGroupCallbacks).

Fields
credential_bundle: &'a CredentialBundlemanaged_group_config: ManagedGroupConfiggroup: MlsGrouppending_proposals: Vec<MLSPlaintext>own_kpbs: Vec<KeyPackageBundle>aad: Vec<u8>resumption_secret_store: ResumptionSecretStoreactive: bool
Implementations

impl<'a> ManagedGroup<'a>

pub fn new(
credential_bundle: &'a CredentialBundle,
managed_group_config: &ManagedGroupConfig,
group_id: GroupId,
key_package_bundle: KeyPackageBundle
) -> Result<Self, ManagedGroupError>

Creates a new group from scratch with only the creator as a member.

pub fn new_from_welcome(
credential_bundle: &'a CredentialBundle,
managed_group_config: &ManagedGroupConfig,
welcome: Welcome,
ratchet_tree: Option<Vec<Option<Node>>>,
key_package_bundle: KeyPackageBundle
) -> Result<Self, GroupError>

Creates a new group from a Welcome message

pub fn add_members(
&mut self,
key_packages: &[KeyPackage]
) -> Result<(Vec<MLSMessage>, Welcome), ManagedGroupError>

Adds members to the group

New members are added by providing a KeyPackage for each member.

If successful, it returns a Vec of MLSMessage and a Welcome message.

pub fn remove_members(
&mut self,
members: &[usize]
) -> Result<(Vec<MLSMessage>, Option<Welcome>), ManagedGroupError>

Removes members from the group

Members are removed by providing the index of their leaf in the tree.

If successful, it returns a Vec of MLSMessage and an optional Welcome message if there were add proposals in the queue of pending proposals.

pub fn propose_add_members(
&mut self,
key_packages: &[KeyPackage]
) -> Result<Vec<MLSMessage>, ManagedGroupError>

Creates proposals to add members to the group

pub fn propose_remove_members(
&mut self,
members: &[usize]
) -> Result<Vec<MLSMessage>, ManagedGroupError>

Creates proposals to remove members from the group

pub fn leave_group(&mut self) -> Result<Vec<MLSMessage>, ManagedGroupError>

Leave the group

pub fn members(&self) -> Vec<Credential>

Gets the current list of members

pub fn process_messages(
&mut self,
messages: Vec<MLSMessage>
) -> Result<Vec<GroupEvent>, ManagedGroupError>

Processes any incoming messages from the DS (MLSPlaintext & MLSCiphertext) and triggers the corresponding callback functions. Return a list of GroupEvent that contain the individual events that occurred while processing messages.

pub fn create_message(
&mut self,
message: &[u8]
) -> Result<MLSMessage, ManagedGroupError>

Creates an application message. Returns ManagedGroupError::UseAfterEviction(UseAfterEviction::Error) if the member is no longer part of the group. Returns ManagedGroupError::PendingProposalsExist if pending proposals exist. In that case .process_pending_proposals() must be called first and incoming messages from the DS must be processed afterwards.

pub fn process_pending_proposals(
&mut self
) -> Result<(Vec<MLSMessage>, Option<Welcome>), ManagedGroupError>

Process pending proposals

pub fn export_secret(
&self,
label: &str,
context: &[u8],
key_length: usize
) -> Result<Vec<u8>, ManagedGroupError>

Exports a secret from the current epoch

pub fn authentication_secret(&self) -> Vec<u8>

Returns the authentication secret

pub fn get_resumption_secret(
&self,
epoch: GroupEpoch
) -> Option<&ResumptionSecret>

Returns a resumption secret for a given epoch. If no resumption secret is available None is returned.

pub fn configuration(&self) -> &ManagedGroupConfig

Gets the configuration

pub fn set_configuration(&mut self, managed_group_config: &ManagedGroupConfig)

Sets the configuration

pub fn aad(&self) -> &[u8]

Gets the AAD used in the framing

pub fn set_aad(&mut self, aad: &[u8])

Sets the AAD used in the framing

pub fn ciphersuite(&self) -> &Ciphersuite

Returns the group's ciphersuite

pub fn is_active(&self) -> bool

Returns whether the own client is still a member of the group or if it was already evicted

pub fn set_credential_bundle(&mut self, credential_bundle: &'a CredentialBundle)

Sets a different CredentialBundle

pub fn credential(&self) -> &Credential

Returns own credential

pub fn group_id(&self) -> &GroupId

Get group ID

pub fn self_update(
&mut self,
key_package_bundle_option: Option<KeyPackageBundle>
) -> Result<(Vec<MLSMessage>, Option<Welcome>), ManagedGroupError>

Updates the own leaf node

A KeyPackageBundle can optionally be provided. If not, a new one will be created on the fly.

If successful, it returns a Vec of MLSMessage and an optional Welcome message if there were add proposals in the queue of pending proposals.

pub fn propose_self_update(
&mut self,
key_package_bundle_option: Option<KeyPackageBundle>
) -> Result<Vec<MLSMessage>, ManagedGroupError>

Creates a proposal to update the own leaf node

pub fn pending_proposals(&self) -> &[MLSPlaintext]

Returns a list of proposal

pub fn load<R: Read>(
reader: R,
credential_bundle: &'a CredentialBundle,
callbacks: &ManagedGroupCallbacks
) -> Result<ManagedGroup<'a>, Error>

Loads the state from persisted state

pub fn save<W: Write>(&self, writer: &mut W) -> Result<(), Error>

Persists the state

pub fn export_ratchet_tree(&self) -> Vec<Option<Node>>

Export the Ratchet Tree

impl<'a> ManagedGroup<'a>

fn plaintext_to_mls_messages(
&mut self,
plaintext_messages: Vec<MLSPlaintext>
) -> Result<Vec<MLSMessage>, ManagedGroupError>

Converts MLSPlaintext to MLSMessage. Depending on whether handshake message should be encrypted, MLSPlaintext messages are encrypted to MLSCiphertext first.

fn validate_proposal(
&self,
proposal: &Proposal,
sender: &LeafIndex,
indexed_members: &HashMap<LeafIndex, Credential>
) -> bool

Validate all pending proposals. The function returns true only if all proposals are valid.

fn validate_inline_proposals(
&self,
proposals: &[ProposalOrRef],
sender: &LeafIndex,
indexed_members: &HashMap<LeafIndex, Credential>
) -> bool

Validates the inline proposals from a Commit message

fn prepare_events(
&self,
ciphersuite: &Ciphersuite,
proposals: &[ProposalOrRef],
sender: LeafIndex,
indexed_members: &HashMap<LeafIndex, Credential>
) -> Vec<GroupEvent>

Prepare the corresponding events for the proposals covered by the Commit

fn prepare_proposal_event(
&self,
proposal: &Proposal,
sender: LeafIndex,
indexed_members: &HashMap<LeafIndex, Credential>
) -> GroupEvent

Prepare the corresponding events for the pending proposal list.

fn auto_save(&self)

Auto-save function

fn indexed_members(&self) -> HashMap<LeafIndex, Credential>

Return a list (LeafIndex, Credential)