Trait openmls::prelude::SerializeBytes

pub trait SerializeBytes: Size {
    // Required method
    fn tls_serialize(&self) -> Result<Vec<u8>, Error>;
}

The SerializeBytes trait provides a function to serialize a struct or enum.

The trait provides one function:

tls_serialize that returns a byte vector

Required Methods

fn tls_serialize(&self) -> Result<Vec<u8>, Error>

Serialize self and return it as a byte vector.