Trait openmls::prelude::Size

pub trait Size {
    // Required method
    fn tls_serialized_len(&self) -> usize;
}

The Size trait needs to be implemented by any struct that should be efficiently serialized. This allows to collect the length of a serialized structure before allocating memory.

Required Methods

fn tls_serialized_len(&self) -> usize