Trait openmls::prelude::TlsDeserializeTrait

pub trait TlsDeserializeTrait: Size {
    fn tls_deserialize<R>(bytes: &mut R) -> Result<Self, Error>
    where
        R: Read,
        Self: Sized;
}

The Deserialize trait defines functions to deserialize a byte slice to a struct or enum.

Required Methods

fn tls_deserialize<R>(bytes: &mut R) -> Result<Self, Error>where
R: Read,
Self: Sized,

This function deserializes the bytes from the provided a std::io::Read and returns the populated struct.

In order to get the amount of bytes read, use Size::tls_serialized_len.