Trait openmls::prelude::TlsSizeTrait

pub trait TlsSizeTrait {
    pub fn tls_serialized_len(&self) -> usize;
}

The Size trait needs to be implemented by any struct that should be efficiently serialized. This allows to collect the length of a serialized structure before allocating memory.

Required methods

pub fn tls_serialized_len(&self) -> usize

Loading content...