Module openmls::schedule

Key schedule

This module contains the types and implementations for key schedule operations. It exposes the EpochAuthenticator & ResumptionPskSecret.

Re-exports

pub use psk::ExternalPsk;

pub use psk::PreSharedKeyId;

pub use psk::Psk;

Modules

errors

Key schedule errors.

psk

Preshared keys.

Structs

EpochAuthenticator

A secret that can be used among members to make sure everyone has the same group state.

ResumptionPskSecret

A group secret that can be used among members to prove that a member was part of a group in a given epoch.