openmls

Module storage

Available on 32-bit or 64-bit only.

OpenMLS Storage

This module serves two purposes:

It implements the Key, Entity and type traits from openmls_traits::storage::traits.

It defines traits that specialize the Storage and Provider traits from openmls_traits. This way, the Rust compiler knows that the concrete types match when we use the Provider in the code.

Traits

OpenMlsProvider

A convenience trait for the OpenMLS provider that defines the storage provider for the current version of storage. Throughout the code, this one should be used instead of openmls_traits::OpenMlsProvider.

PublicStorageProvider

A convenience trait for the current version of the public storage. Throughout the code, this one should be used instead of openmls_traits::public_storage::PublicStorageProvider.

StorageProvider

A convenience trait for the current version of the storage. Throughout the code, this one should be used instead of openmls_traits::storage::StorageProvider.