Struct openmls::test_utils::test_framework::client::Client

pub struct Client {
    pub identity: Vec<u8>,
    pub credentials: HashMap<CiphersuiteName, Credential>,
    pub key_store: KeyStore,
    pub groups: RefCell<HashMap<GroupId, ManagedGroup>>,
}

The client contains the necessary state for a client in the context of MLS. It contains the group states, as well as a reference to a KeyStore containing its CredentialBundles. The key_package_bundles field contains generated KeyPackageBundles that are waiting to be used for new groups.

Fields
identity: Vec<u8>

Name of the client.

credentials: HashMap<CiphersuiteName, Credential>

Ciphersuites supported by the client.

key_store: KeyStoregroups: RefCell<HashMap<GroupId, ManagedGroup>>
Implementations

impl Client

pub fn get_fresh_key_package(
&self,
ciphersuites: &[CiphersuiteName]
) -> Result<KeyPackage, ClientError>

Generate a fresh key package bundle and store it in self.key_package_bundles. The first ciphersuite determines the credential used to generate the KeyPackageBundle. Returns the corresponding KeyPackage.

pub fn create_group(
&self,
group_id: GroupId,
managed_group_config: ManagedGroupConfig,
ciphersuite: &Ciphersuite
) -> Result<(), ClientError>

Create a group with the given group_id, ciphersuite and managed_group_config. Throws an error if the client doesn’t support the ciphersuite, i.e. if no corresponding CredentialBundle exists.

pub fn join_group(
&self,
managed_group_config: ManagedGroupConfig,
welcome: Welcome,
ratchet_tree: Option<Vec<Option<Node>>>
) -> Result<(), ClientError>

Join a group based on the given welcome and ratchet_tree. The group is created with the given ManagedGroupConfig. Throws an error if no KeyPackage exists matching the Welcome, if the client doesn’t support the ciphersuite, or if an error occurs processing the Welcome.

pub fn receive_messages_for_group(
&self,
message: &MlsMessage
) -> Result<(), ClientError>

Have the client process the given messages. Returns an error if an error occurs during message processing or if no group exists for one of the messages.

pub fn get_members_of_group(
&self,
group_id: &GroupId
) -> Result<Vec<(usize, Credential)>, ClientError>

Get the credential and the index of each group member of the group with the given id. Returns an error if no group exists with the given group id.

pub fn self_update(
&self,
action_type: ActionType,
group_id: &GroupId,
key_package_bundle_option: Option<KeyPackageBundle>
) -> Result<(MlsMessage, Option<Welcome>), ClientError>

Have the client either propose or commit (depending on the action_type) a self update in the group with the given group id. Optionally, a KeyPackageBundle can be provided, which the client will update their leaf with. Returns an error if no group with the given group id can be found or if an error occurs while creating the update.

pub fn add_members(
&self,
action_type: ActionType,
group_id: &GroupId,
key_packages: &[KeyPackage]
) -> Result<(Vec<MlsMessage>, Option<Welcome>), ClientError>

Have the client either propose or commit (depending on the action_type) adding the clients with the given KeyPackages to the group with the given group id. Returns an error if no group with the given group id can be found or if an error occurs while performing the add operation.

pub fn remove_members(
&self,
action_type: ActionType,
group_id: &GroupId,
target_indices: &[usize]
) -> Result<(Vec<MlsMessage>, Option<Welcome>), ClientError>

Have the client either propose or commit (depending on the action_type) removing the clients with the given indices from the group with the given group id. Returns an error if no group with the given group id can be found or if an error occurs while performing the remove operation.