Module openmls::treesync

This module implements the ratchet tree component of MLS.

About

This module provides the TreeSync struct, which contains the state shared between a group of MLS clients in the shape of a tree, where each non-blank leaf corresponds to one group member. The functions provided by its implementation allow the creation of a [TreeSyncDiff] instance, which in turn can be mutably operated on and merged back into the original TreeSync instance.

The submodules of this module define the nodes of the tree (nodes), helper functions and structs for the algorithms used to sync the tree across the group ([hashes]) and the diff functionality ([diff]).

Finally, this module contains the [treekem] module, which allows the encryption and decryption of updates to the tree.

Don’t Panic!

Functions in this module should never panic. However, if there is a bug in the implementation, a function will return an unrecoverable LibraryError. This means that some functions that are not expected to fail and throw an error, will still return a Result since they may throw a LibraryError.

Modules

node

This module contains types and methods around the Node enum. The variants of the enum are [LeafNode] and [ParentNode], both of which are defined in the respective [leaf_node] and [parent_node] submodules.

Structs

TreeSync

The TreeSync struct holds an [MlsBinaryTree] instance, which contains the state that is synced across the group, as well as the LeafIndex pointing to the leaf of this group member and the current hash of the tree.

Enums

TreeSyncError

Type Definitions

LeafIndex